Added new admin floors API
