--- src/api/engineering/FacilitiesAPI.jsx
@@ -1,53 +1,0 @@
-const FacilitiesData = {
-    floors: [
-      {
-        floor: 1,
-        rooms: [
-          {
-            roomNumber: "1",
-            name: "Fab-Lab",
-            description: "Serves as the front desk for visitors and inquiries. Includes a waiting area and an information counter.",
-            image: "/images/rooms/reception.jpg"
-          },
-          {
-            roomNumber: "2",
-            name: "SAC",
-            description: "A fully equipped meeting room with a projector, whiteboard, and seating for 15 people.",
-            image: "/images/rooms/conference.jpg"
-          },
-          {
-            roomNumber: "3",
-            name: "Old-Canteen",
-            description: "Office space for the IT department, equipped with multiple workstations and a server cabinet.",
-            image: "/images/rooms/it_office.jpg"
-          },
-          {
-            roomNumber: "4",
-            name: "Bistro Canteen",
-            description: "Secured space for storing office supplies, equipment, and documents.",
-            image: "/images/rooms/storage.jpg"
-          },
-          {
-            roomNumber: "5",
-            name: "Centrum",
-            description: "Secured space for storing office supplies, equipment, and documents.",
-            image: "/images/rooms/storage.jpg"
-          }
-        ],
-        bathrooms: [
-          {
-            bathroomNumber: "B1",
-            description: "Men's bathroom with two stalls and a sink.",
-            image: "/images/rooms/bathroom_men.jpg"
-          },
-          {
-            bathroomNumber: "B2",
-            description: "Women's bathroom with two stalls and a sink.",
-            image: "/images/rooms/bathroom_women.jpg"
-          }
-        ]
-      }
-    ]
-  };
-
-export default  FacilitiesData;
