--- src/api/new admin/floors/NewAdminFloor1.jsx
@@ -1,294 +1,0 @@
-import React from 'react'
-import { useContext, useState } from "react";
-
-export const NewAdminFloor1 = () => {
-
-    const [showInfoPanel, setShowInfoPanel] = useState(false);
-    const [roomInfo, setRoomInfo] = useState(null);
-    const [emoji, setEmoji] = useState("👷‍♂️");
-
-    const Floor1 = {
-    "1": {
-    "rooms": {
-        
-        }
-    }
-
-  }
-
-    const roomClicked = (key) =>{
-    setRoomInfo({ key, ...Floor1["1"].rooms[key] });
-    console.log(roomInfo);
-  } 
-
-    const roomOfs = -5;
-
-    const roomOnMouseEnter = (e, room, roomText) =>{
-    e.currentTarget.setAttribute('fill','#FFFFFF');
-    e.currentTarget.style.transform = `translate(${roomOfs}px, ${roomOfs}px)`; 
-    document.querySelector(roomText).style.transform = `translate(${roomOfs}px, ${roomOfs}px)`;
-    const roomStroke = document.querySelectorAll(room); 
-    console.log(roomStroke)
-    roomStroke.forEach(ro => {
-            ro.style.transform = `translate(${roomOfs}px, ${roomOfs}px)`;
-    });
-  }
-    const roomOnMouseLeave = (e, room, roomText, roomColor) =>{
-    e.currentTarget.setAttribute('fill', roomColor);
-    e.currentTarget.style.transform = 'translate(0px, 0px)'; 
-    document.querySelector(roomText).style.transform = 'translate(0px, 0px)'; 
-    const roomStroke = document.querySelectorAll(room); 
-    roomStroke.forEach(ro => {
-            ro.style.transform = 'translate(0px, 0px)'; 
-    });
-  }
-
-  return (
-    <div className='-mt-6 mr-32 lg:flex w-screen justify-center items-center'>
-        <svg id="layers" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 2590.93 1063.56">
-        <defs>
-            <clipPath id="clippath">
-            <polygon points="636.81 592.2 756.78 319.1 2142.84 538.62 2022.87 811.72 636.81 592.2" fill='none' strokeWidth='0px'/>
-            </clipPath>
-            <clipPath id="clippath-1">
-            <polyline points="733.26 372.65 657.98 544 1136.28 619.76 1117.8 661.82 1247.64 682.39 1244.95 688.51 1413.63 715.23 1434.8 667.04 2044.04 763.53 2118.97 592.94 1948.08 565.87 1953.46 553.63 1830.27 534.12 1824.9 546.36 1691.73 525.27 1715.59 470.96 1184.03 386.76 1159.83 441.84 1026.67 420.75 1033.39 405.45 1028.51 404.68 919.65 387.44 914.65 386.65 908.6 400.42 733.26 372.65" fill="none" strokeWidth="0px"/>
-            </clipPath>
-        </defs>
-        <path d="M1701.44,347.28h378.9v278.38c0,97.04-78.78,175.82-175.82,175.82h-203.08c-27.42,0-49.68-22.26-49.68-49.68v-354.83c0-27.42,22.26-49.68,49.68-49.68Z" fill= "lime" opacity= ".21" strokeWidth= "0px"/>
-        <path d="M835.17,334.08h327.78c28.12,0,50.94,22.83,50.94,50.94v284.69c0,65.44-53.13,118.56-118.56,118.56h-315.26v-399.09c0-30.41,24.69-55.11,55.11-55.11Z" fill= "#f47cf7" opacity= ".21" strokeWidth= "0px"/>
-        <path d="M1653.26,441.79l-468.3-156.5c-17.4-5.82-34.65,9.21-31.27,27.25l88.48,471.99c3.38,18.03,24.9,25.8,39.01,14.07l379.82-315.49c14.51-12.05,10.14-35.34-7.75-41.32Z" fill= "#ff0" opacity= ".19" strokeWidth= "0px"/>
-        <circle cx="1565.75" cy="669.93" r="130.57" fill="#ef507a" opacity=".14" strokeWidth="0px"/>
-        <g id="floor">
-            <g isolation="isolate">
-            <polygon points="914.41 387.18 914.13 401.36 908.09 415.12 908.36 400.94 914.41 387.18" fill="#291455" strokeWidth="0px"/>
-            <polygon points="908.36 400.94 908.09 415.12 732.75 387.35 733.02 373.17 908.36 400.94" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1032.35 405.86 1032.08 420.04 914.13 401.36 914.41 387.18 1032.35 405.86" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1032.36 405.86 1032.08 420.04 1025.36 435.34 1025.64 421.16 1032.36 405.86" fill="#767676" strokeWidth="0px"/>
-            <polygon points="1033.35 405.53 1033.07 419.71 1026.35 435.01 1026.63 420.83 1033.35 405.53" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1183.79 387.29 1183.51 401.47 1159.32 456.55 1159.59 442.37 1183.79 387.29" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1159.59 442.37 1159.32 456.55 1025.36 435.33 1025.63 421.15 1159.59 442.37" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1714.56 471.36 1714.28 485.54 1183.51 401.47 1183.79 387.29 1714.56 471.36" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1714.56 471.36 1714.28 485.54 1690.43 539.85 1690.7 525.67 1714.56 471.36" fill="#767676" strokeWidth="0px"/>
-            <polygon points="1715.54 471.03 1715.27 485.21 1691.41 539.53 1691.68 525.35 1715.54 471.03" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1824.66 546.89 1824.38 561.07 1690.43 539.85 1690.7 525.67 1824.66 546.89" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1830.03 534.65 1829.76 548.83 1824.38 561.07 1824.66 546.89 1830.03 534.65" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1952.42 554.04 1952.15 568.22 1829.76 548.83 1830.03 534.65 1952.42 554.04" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1952.43 554.04 1952.15 568.22 1946.78 580.46 1947.05 566.27 1952.43 554.04" fill="#767676" strokeWidth="0px"/>
-            <polygon points="1953.41 553.71 1953.13 567.89 1947.76 580.13 1948.03 565.95 1953.41 553.71" fill="#291455" strokeWidth="0px"/>
-            <polygon points="733.02 373.17 732.75 387.35 657.67 558.26 657.95 544.08 733.02 373.17" fill="#291455" strokeWidth="0px"/>
-            <polygon points="2117.94 593.34 2117.67 607.52 1946.77 580.45 1947.05 566.27 2117.94 593.34" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1136.24 619.83 1135.97 634.01 657.67 558.26 657.95 544.08 1136.24 619.83" fill="#747474" strokeWidth="0px"/>
-            <polygon points="1135.25 620.16 1134.97 634.34 656.68 558.58 656.95 544.4 1135.25 620.16" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1136.24 619.83 1135.96 634.01 1117.48 676.08 1117.76 661.9 1136.24 619.83" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1247.6 682.47 1247.32 696.65 1117.48 676.08 1117.76 661.9 1247.6 682.47" fill="#747474" strokeWidth="0px"/>
-            <polygon points="1246.61 682.79 1246.33 696.97 1116.49 676.41 1116.76 662.23 1246.61 682.79" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1247.6 682.47 1247.32 696.65 1244.63 702.77 1244.91 688.59 1247.6 682.47" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1433.97 666.99 1433.69 681.17 1412.52 729.36 1412.8 715.18 1433.97 666.99" fill="#767676" strokeWidth="0px"/>
-            <polygon points="1412.8 715.18 1412.52 729.36 1244.64 702.77 1244.91 688.59 1412.8 715.18" fill="#747474" strokeWidth="0px"/>
-            <polygon points="2043.21 763.48 2042.93 777.66 1433.69 681.17 1433.97 666.99 2043.21 763.48" fill="#747474" strokeWidth="0px"/>
-            <polygon points="1434.55 667.56 1434.28 681.74 1413.11 729.93 1413.38 715.75 1434.55 667.56" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1413.38 715.75 1413.11 729.93 1243.64 703.09 1243.92 688.91 1413.38 715.75" fill="#281454" strokeWidth="0px"/>
-            <polygon points="2117.95 593.34 2117.67 607.52 2042.93 777.66 2043.21 763.48 2117.95 593.34" fill="#767676" strokeWidth="0px"/>
-            <polygon points="1714.56 471.36 1690.7 525.67 1824.66 546.89 1830.04 534.65 1952.43 554.04 1947.05 566.27 2117.95 593.34 2043.21 763.48 1433.97 666.99 1412.8 715.18 1244.91 688.59 1247.6 682.47 1117.76 661.9 1136.24 619.83 657.95 544.08 733.02 373.17 908.36 400.94 914.41 387.18 1032.36 405.86 1025.64 421.16 1159.6 442.37 1183.79 387.3 1714.56 471.36" fill="#9b9b9b" strokeWidth="0px"/>
-            <polygon points="2043.8 764.06 2043.53 778.24 1434.28 681.74 1434.55 667.56 2043.8 764.06" fill="#281454" strokeWidth="0px"/>
-            <polygon points="2118.93 593.02 2118.66 607.2 2043.53 778.24 2043.8 764.06 2118.93 593.02" fill="#291455" strokeWidth="0px"/>
-            <path d="M1948.03,565.95l170.9,27.07-75.14,171.04-609.25-96.5-21.17,48.19-169.46-26.84,2.69-6.12-129.84-20.56,18.48-42.07-478.29-75.75,75.47-171.81,175.34,27.77,6.05-13.77,119.54,18.93-6.72,15.3,132.37,20.97,24.19-55.08,532.35,84.32-23.86,54.31,132.38,20.97,5.38-12.24,123.97,19.63-5.38,12.24ZM2043.2,763.48l74.74-170.14-170.9-27.07,5.38-12.24-122.39-19.38-5.38,12.24-133.96-21.22,23.86-54.31-530.77-84.07-24.2,55.08-133.96-21.22,6.72-15.3-117.95-18.68-6.05,13.77-175.34-27.77-75.08,170.91,478.29,75.75-18.48,42.07,129.84,20.56-2.69,6.12,167.88,26.59,21.17-48.19,609.24,96.49" fill="#361b70" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="grid" opacity=".17">
-            <g isolation="isolate">
-            <g clip-path="url(#clippath)">
-                <g clip-path="url(#clippath-1)">
-                <path d="M1126.52,657.44l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1183.72,666.5l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1240.93,675.56l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1298.14,684.62l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1355.33,693.68l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1136.03,635.78l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1193.24,644.85l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1250.44,653.91l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1307.65,662.97l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1364.85,672.03l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M687.91,541.65l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M745.11,550.71l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M802.32,559.77l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M859.52,568.83l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M916.73,577.89l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M973.93,586.95l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1031.14,596.01l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1088.34,605.07l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1145.54,614.13l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1202.75,623.19l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1259.95,632.25l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1317.16,641.31l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1374.36,650.37l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1431.57,659.43l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1488.77,668.49l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1545.97,677.55l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1603.18,686.61l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1660.39,695.68l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1717.59,704.74l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1774.79,713.8l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1832,722.86l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1889.2,731.92l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1946.4,740.98l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M697.42,520l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M754.62,529.06l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M811.83,538.12l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M869.04,547.18l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M926.24,556.24l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M983.44,565.3l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1040.65,574.36l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1097.85,583.42l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1155.05,592.48l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1212.26,601.54l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1269.47,610.6l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1326.67,619.66l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1383.87,628.72l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1441.08,637.78l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1498.29,646.84l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1555.48,655.9l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1612.69,664.96l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1669.9,674.02l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1727.1,683.08l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1784.3,692.14l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1841.51,701.2l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1898.72,710.26l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1955.91,719.32l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M706.94,498.34l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M764.14,507.4l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M821.34,516.46l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M878.55,525.52l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M935.76,534.58l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M992.95,543.64l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1050.16,552.7l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1107.37,561.76l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1164.57,570.82l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1221.77,579.88l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1278.98,588.94l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1336.19,598l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1393.39,607.06l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1450.59,616.12l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1507.8,625.18l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1565,634.24l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1622.2,643.3l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1679.41,652.36l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1736.62,661.43l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1793.82,670.48l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1851.02,679.55l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1908.23,688.61l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1965.43,697.67l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M716.45,476.69l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M773.65,485.75l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M830.85,494.81l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M888.06,503.87l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M945.27,512.93l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1002.47,521.99l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1059.67,531.05l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1116.88,540.11l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1174.08,549.17l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1231.29,558.23l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1288.49,567.29l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1345.7,576.35l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1402.9,585.41l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1460.1,594.47l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1517.31,603.53l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1574.51,612.59l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1631.72,621.65l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1688.92,630.71l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1746.13,639.77l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1803.33,648.83l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1860.53,657.89l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1917.74,666.95l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1974.94,676.01l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M725.96,455.03l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M783.16,464.09l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M840.37,473.15l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M897.57,482.21l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M954.78,491.28l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1011.98,500.33l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1069.19,509.4l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1126.39,518.46l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1183.59,527.52l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1240.8,536.58l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1298,545.64l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1355.21,554.7l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1412.41,563.76l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1469.62,572.82l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1526.82,581.88l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1584.02,590.94l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1641.23,600l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1698.43,609.06l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1755.64,618.12l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1812.84,627.18l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1870.05,636.24l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1927.25,645.3l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1984.45,654.36l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M735.48,433.38l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M792.67,442.44l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M849.88,451.5l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M907.09,460.56l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M964.29,469.62l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1021.49,478.68l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1078.7,487.74l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1135.91,496.8l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1193.1,505.86l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1250.31,514.92l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1307.52,523.98l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1364.72,533.04l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1421.92,542.1l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1479.13,551.16l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1536.34,560.22l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1593.53,569.28l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1650.74,578.34l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1707.95,587.4l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1765.16,596.46l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1822.35,605.52l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1879.56,614.58l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1936.77,623.64l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1993.97,632.7l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M744.99,411.72l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M802.19,420.78l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M859.39,429.84l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M916.6,438.9l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M973.81,447.96l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1031,457.02l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1088.21,466.08l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1145.42,475.15l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1202.62,484.2l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1259.82,493.27l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1317.03,502.33l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1374.24,511.39l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1431.43,520.45l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1488.64,529.51l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1545.85,538.57l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1603.05,547.63l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1660.25,556.69l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1717.46,565.75l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1774.67,574.81l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1831.86,583.87l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1889.07,592.93l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1946.28,601.99l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M2003.48,611.05l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M926.11,417.25l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1154.93,453.49l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1212.13,462.55l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1269.34,471.61l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1326.54,480.67l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1383.75,489.73l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1440.95,498.79l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1498.16,507.85l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1555.36,516.91l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1612.56,525.97l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1841.38,562.21l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1164.44,431.83l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1221.64,440.89l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1278.85,449.95l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1336.06,459.02l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1393.26,468.08l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1450.46,477.14l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1507.67,486.2l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1564.87,495.26l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1622.07,504.32l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1173.96,410.18l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1231.15,419.24l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1288.36,428.3l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1345.57,437.36l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1402.77,446.42l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1459.97,455.48l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1517.18,464.54l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1574.39,473.6l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1631.58,482.66l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1699.06,468.34l-.97-.15-.76,1.73-56.24-8.91.76-1.73-.96-.15-.76,1.73-56.24-8.91.76-1.73-.97-.15-.76,1.73-56.24-8.91.76-1.73-.97-.15-.76,1.73-56.24-8.91.76-1.73-.97-.15-.76,1.73-56.23-8.91.76-1.73-.97-.15-.76,1.73-56.24-8.91.76-1.73-.97-.15-.76,1.73-56.24-8.91.76-1.73-.97-.15-.76,1.73-56.24-8.91.76-1.73-.96-.15-.76,1.73-56.24-8.91.76-1.73-.2-.03-24.2,55.08-.77-.12-5.1,11.61-56.24-8.91,5.1-11.61-.97-.15-5.1,11.61-56.24-8.91,5.1-11.61-.97-.15-5.1,11.61-56.23-8.91,9.27-21.1,38.25,6.06.24-.56-38.25-6.06,2.31-5.26-.97-.15-2.31,5.26-56.24-8.91,2.31-5.26-.97-.15-2.31,5.26-22.32-3.54-.24.56,22.32,3.54-9.27,21.1-56.24-8.91,5.77-13.14-.97-.15-5.77,13.14-56.24-8.91,5.77-13.14-.96-.15-5.77,13.14-56.24-8.91,5.77-13.14-.97-.15-5.77,13.14-26.05-4.13-.24.56,26.05,4.13-9.27,21.1-26.05-4.13-.24.56,26.05,4.13-9.27,21.1-26.05-4.13-.24.56,26.05,4.13-9.27,21.1-26.05-4.13-.24.55,26.05,4.13-9.27,21.1-26.05-4.13-.24.56,26.05,4.13-9.27,21.1-26.05-4.13-.24.56,26.05,4.13-9.27,21.1-26.05-4.13-.24.55,26.05,4.13-9.27,21.1-26.05-4.13-.24.56,26.05,4.13-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.96.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.23,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-9.27,21.1-5.39-.85-.24.56,5.39.85-9.27,21.1-5.39-.85-.24.56,5.39.85-2.12,4.83.96.15,2.12-4.83,56.24,8.91-2.12,4.83.97.15,2.12-4.83,56.24,8.91-2.12,4.83.97.15,2.12-4.83,56.24,8.91-4.81,10.96.97.15,4.81-10.96,56.23,8.91-4.81,10.96.97.15,4.81-10.96,56.24,8.91-4.81,10.96.97.15,4.81-10.96,6.14.97.24-.56-6.14-.97,9.27-21.1,6.14.97.24-.56-6.14-.97,9.27-21.1,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.96.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.23,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,56.23,8.91-2.67,6.07.97.15,2.67-6.07,56.24,8.91-2.67,6.07.97.15,2.67-6.07,43.34,6.86.24-.56-43.34-6.86,9.27-21.1,43.34,6.86.24-.55-43.34-6.86,9.27-21.1,43.34,6.86.24-.56-43.34-6.86,9.27-21.1,43.34,6.86.24-.56-43.34-6.86,9.27-21.1,43.34,6.86.24-.55-43.34-6.86,9.27-21.1,43.34,6.86.24-.56-43.34-6.86,9.27-21.1,43.34,6.86.24-.56-43.34-6.86,9.27-21.1,43.34,6.86.24-.56-43.34-6.86,5.44-12.37-.97-.15-5.44,12.37-56.24-8.91,5.44-12.37-.97-.15-5.44,12.37-56.23-8.91,5.44-12.37-.97-.15-5.44,12.37-56.24-8.91,9.27-21.1,44.06,6.98.24-.56-44.06-6.98,1.3-2.96-.97-.15-1.3,2.96-56.24-8.91,1.3-2.96-.97-.15-1.3,2.96-20.95-3.32-.24.56,20.95,3.32-9.27,21.1-56.23-8.91,5.44-12.37-.97-.15-5.44,12.37-56.24-8.91,5.44-12.37-.97-.15-5.44,12.37-56.24-8.91,9.27-21.1,16.53,2.62.24-.56-16.53-2.62,9.27-21.1,16.53,2.62.24-.56-16.53-2.62,9.27-21.1,16.53,2.62.24-.56-16.53-2.62.76-1.73" fill="#fff" strokeWidth="0px"/>
-                </g>
-            </g>
-            </g>
-        </g>
-        <g id="long_stairs" data-name="long stairs">
-            <g isolation="isolate">
-            <polygon points="1211.78 418.6 1211.71 421.92 1207.99 430.41 1208.05 427.08 1211.78 418.6" fill="#979797" strokeWidth="0px"/>
-            <polygon points="1662.56 490 1662.5 493.32 1211.71 421.92 1211.78 418.6 1662.56 490" fill="#959595" strokeWidth="0px"/>
-            <polygon points="1212.37 419.18 1212.31 422.5 1208.98 430.08 1209.04 426.76 1212.37 419.18" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1188.87 392.95 1188.81 396.28 1168.07 443.49 1168.13 440.17 1188.87 392.95" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1357.28 450.72 1357.22 454.04 1207.99 430.41 1208.05 427.08 1357.28 450.72" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1702.44 474.29 1702.38 477.62 1188.81 396.28 1188.87 392.95 1702.44 474.29" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1357.28 450.72 1357.22 454.04 1349.24 472.19 1349.31 468.87 1357.28 450.72" fill="#979797" strokeWidth="0px"/>
-            <polygon points="1349.31 468.87 1349.24 472.19 1168.07 443.49 1168.13 440.17 1349.31 468.87" fill="#959595" strokeWidth="0px"/>
-            <polygon points="1349.9 469.44 1349.83 472.76 1167.09 443.82 1167.15 440.5 1349.9 469.44" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1358.26 450.39 1358.2 453.71 1349.83 472.76 1349.9 469.44 1358.26 450.39" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1661.58 490.32 1661.52 493.65 1212.31 422.5 1212.37 419.18 1661.58 490.32" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1508.84 473.9 1508.77 477.23 1500.64 495.73 1500.71 492.41 1508.84 473.9" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1659.17 497.71 1659.11 501.04 1508.77 477.23 1508.84 473.9 1659.17 497.71" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1662.56 490 1662.5 493.32 1659.11 501.04 1659.17 497.71 1662.56 490" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1681.88 521.1 1681.82 524.42 1500.64 495.73 1500.71 492.41 1681.88 521.1" fill="#959595" strokeWidth="0px"/>
-            <polygon points="1702.44 474.29 1702.38 477.62 1681.82 524.42 1681.88 521.1 1702.44 474.29" fill="#979797" strokeWidth="0px"/>
-            <path d="M1188.29,392.38l515.15,81.59-20.96,47.71-182.75-28.95,8.53-19.41,150.34,23.81,2.99-6.82-449.21-71.15-3.33,7.58,149.22,23.63-8.37,19.05-182.75-28.94,21.14-48.12ZM1681.88,521.1l20.56-46.81-513.57-81.34-20.74,47.22,181.18,28.7,7.97-18.15-149.23-23.64,3.73-8.48,450.79,71.4-3.39,7.72-150.34-23.81-8.13,18.5,181.18,28.7" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="1702.44 474.29 1681.88 521.1 1500.71 492.41 1508.84 473.9 1659.17 497.71 1662.56 490 1211.78 418.6 1208.05 427.08 1357.28 450.72 1349.31 468.87 1168.13 440.17 1188.87 392.95 1702.44 474.29" fill="#c7c7c7" strokeWidth="0px"/>
-            <polygon points="1703.44 473.97 1703.37 477.29 1682.41 525.01 1682.48 521.68 1703.44 473.97" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1682.48 521.68 1682.41 525.01 1499.66 496.06 1499.72 492.74 1682.48 521.68" fill="#281454" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_1" data-name="1">
-            <g isolation="isolate">
-            <polygon points="845.81 382.52 845.54 396.7 730.91 378.54 731.19 364.36 845.81 382.52" fill="#281454" strokeWidth="0px"/>
-            <polygon points="845.81 382.52 845.54 396.7 829.74 432.65 830.02 418.47 845.81 382.52" fill="#5c8229" strokeWidth="0px"/>
-            <polygon points="846.8 382.19 846.52 396.37 830.73 432.32 831 418.14 846.8 382.19" fill="#291455" strokeWidth="0px"/>
-            <polygon points="863.31 423.74 863.03 437.92 829.74 432.65 830.02 418.47 863.31 423.74" fill="#281454" strokeWidth="0px"/>
-            <polygon points="731.19 364.36 730.91 378.54 657.86 544.85 658.13 530.67 731.19 364.36" fill="#291455" strokeWidth="0px"/>
-            <polygon points="806.04 554.1 805.77 568.28 657.86 544.85 658.13 530.67 806.04 554.1" fill="#5b8129" strokeWidth="0px"/>
-            <polygon points="863.31 423.74 863.03 437.92 805.77 568.28 806.04 554.1 863.31 423.74" fill="#5c8229" strokeWidth="0px"/>
-            <path d="M831,418.14l33.3,5.27-57.66,131.26-149.5-23.68,73.46-167.22,116.2,18.4-15.79,35.95ZM806.04,554.1l57.26-130.35-33.29-5.27,15.79-35.95-114.62-18.15-73.06,166.31,147.91,23.43" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="845.81 382.52 830.02 418.47 863.31 423.74 806.04 554.1 658.13 530.67 731.19 364.36 845.81 382.52" fill="#79ab36" strokeWidth="0px"/>
-            <polygon points="864.3 423.41 864.03 437.59 806.36 568.86 806.64 554.68 864.3 423.41" fill="#291455" strokeWidth="0px"/>
-            <polygon points="806.64 554.68 806.36 568.86 656.86 545.18 657.14 531 806.64 554.68" fill="#281454" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_2" data-name="2">
-            <g isolation="isolate">
-            <polygon points="980.27 442.64 979.99 456.82 864.26 438.49 864.53 424.31 980.27 442.64" fill="#281454" strokeWidth="0px"/>
-            <polygon points="864.53 424.31 864.26 438.49 807 568.85 807.27 554.67 864.53 424.31" fill="#291455" strokeWidth="0px"/>
-            <polygon points="923 573 922.73 587.18 807 568.85 807.27 554.67 923 573" fill="#a37226" strokeWidth="0px"/>
-            <polygon points="980.27 442.64 979.99 456.82 922.73 587.18 923 573 980.27 442.64" fill="#a57427" strokeWidth="0px"/>
-            <polygon points="980.27 442.64 923 573 807.27 554.67 864.53 424.31 980.27 442.64" fill="#d99833" strokeWidth="0px"/>
-            <polygon points="923.59 573.57 923.31 587.75 806 569.17 806.28 554.99 923.59 573.57" fill="#281454" strokeWidth="0px"/>
-            <polygon points="981.25 442.31 980.98 456.49 923.31 587.75 923.59 573.57 981.25 442.31" fill="#291455" strokeWidth="0px"/>
-            <path d="M863.94,423.73l117.31,18.58-57.66,131.26-117.31-18.58,57.66-131.26ZM923,573l57.26-130.36-115.73-18.33-57.26,130.36,115.73,18.33" fill="#361b70" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_3" data-name="3">
-            <g isolation="isolate">
-            <polygon points="1250.28 485.22 1250.01 499.4 982.25 456.99 982.52 442.81 1250.28 485.22" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1250.28 485.22 1250.01 499.4 1249.78 499.93 1250.05 485.75 1250.28 485.22" fill="#9d3550" strokeWidth="0px"/>
-            <polygon points="1251.27 484.89 1251 499.07 1250.57 500.06 1250.84 485.88 1251.27 484.89" fill="#291455" strokeWidth="0px"/>
-            <polygon points="982.52 442.81 982.25 456.99 924.98 587.35 925.26 573.17 982.52 442.81" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1191.91 615.4 1191.64 629.58 924.98 587.35 925.26 573.17 1191.91 615.4" fill="#9b344f" strokeWidth="0px"/>
-            <polygon points="1250.05 485.75 1249.78 499.93 1191.64 629.58 1191.91 615.4 1250.05 485.75" fill="#9d3550" strokeWidth="0px"/>
-            <path d="M981.92,442.23l269.35,42.66-.43.98-58.35,130.11-268.23-42.48,57.66-131.26ZM1250.05,485.75l.23-.53-267.76-42.41-57.26,130.36,266.65,42.23,58.14-129.65" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="1250.28 485.22 1250.05 485.75 1191.91 615.4 925.26 573.17 982.52 442.81 1250.28 485.22" fill="#cf4569" strokeWidth="0px"/>
-            <polygon points="1192.5 615.98 1192.22 630.16 923.99 587.68 924.26 573.5 1192.5 615.98" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1250.84 485.88 1250.57 500.06 1192.22 630.16 1192.5 615.98 1250.84 485.88" fill="#291455" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="pwd_stairs" data-name="pwd stairs">
-            <g isolation="isolate">
-            <polygon points="1264.33 639.3 1264.29 641.29 1136.17 621 1136.2 619.01 1264.33 639.3" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1255.79 658.74 1255.75 660.73 1154.58 644.71 1154.62 642.72 1255.79 658.74" fill="#959595" strokeWidth="0px"/>
-            <polygon points="1155.22 643.29 1155.18 645.29 1154.53 646.75 1154.57 644.75 1155.22 643.29" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1154.62 642.72 1154.58 644.71 1153.54 647.07 1153.58 645.08 1154.62 642.72" fill="#979797" strokeWidth="0px"/>
-            <polygon points="1264.33 639.3 1264.29 641.29 1255.75 660.73 1255.79 658.74 1264.33 639.3" fill="#979797" strokeWidth="0px"/>
-            <polygon points="1265.31 638.97 1265.27 640.97 1256.34 661.31 1256.37 659.31 1265.31 638.97" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1256.37 659.31 1256.34 661.31 1155.18 645.29 1155.22 643.29 1256.37 659.31" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1254.75 661.1 1254.71 663.1 1153.54 647.07 1153.58 645.08 1254.75 661.1" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1136.2 619.01 1136.17 621 1118.22 661.86 1118.26 659.86 1136.2 619.01" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1254.75 661.1 1254.71 663.1 1246.34 682.15 1246.38 680.16 1254.75 661.1" fill="#979797" strokeWidth="0px"/>
-            <polygon points="1246.38 680.16 1246.34 682.15 1118.22 661.86 1118.26 659.86 1246.38 680.16" fill="#959595" strokeWidth="0px"/>
-            <polygon points="1264.33 639.3 1255.79 658.74 1154.62 642.72 1153.58 645.08 1254.75 661.1 1246.38 680.16 1118.26 659.86 1136.2 619.01 1264.33 639.3" fill="#c7c7c7" strokeWidth="0px"/>
-            <path d="M1256.37,659.31l-101.16-16.02-.64,1.46,101.16,16.02-8.77,19.96-129.69-20.54,18.35-41.76,129.69,20.54-8.94,20.34ZM1254.75,661.1l-101.17-16.02,1.04-2.36,101.17,16.02,8.54-19.44-128.12-20.29-17.95,40.86,128.12,20.29,8.37-19.05" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="1246.96 680.74 1246.93 682.73 1117.23 662.19 1117.27 660.2 1246.96 680.74" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1255.73 660.78 1255.69 662.77 1246.93 682.73 1246.96 680.74 1255.73 660.78" fill="#291455" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_4" data-name="4">
-            <g isolation="isolate">
-            <polygon points="1520.98 634.13 1520.72 647.76 1449.46 636.47 1449.72 622.85 1520.98 634.13" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1449.72 622.85 1449.46 636.47 1435.61 668 1435.87 654.37 1449.72 622.85" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1507.13 665.66 1506.87 679.29 1435.61 668 1435.87 654.37 1507.13 665.66" fill="#00a600" strokeWidth="0px"/>
-            <polygon points="1520.98 634.13 1520.72 647.76 1506.87 679.29 1507.13 665.66 1520.98 634.13" fill="#00a700" strokeWidth="0px"/>
-            <polygon points="1520.98 634.13 1507.13 665.66 1435.87 654.37 1449.72 622.85 1520.98 634.13" fill="#00dc00" strokeWidth="0px"/>
-            <polygon points="1507.58 666.1 1507.32 679.72 1434.87 668.25 1435.13 654.62 1507.58 666.1" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1521.73 633.89 1521.47 647.51 1507.32 679.72 1507.58 666.1 1521.73 633.89" fill="#291455" strokeWidth="0px"/>
-            <path d="M1449.28,622.41l72.45,11.48-14.15,32.21-72.45-11.48,14.15-32.21ZM1507.13,665.66l13.85-31.53-71.26-11.29-13.85,31.53,71.26,11.29" fill="#361b70" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_5" data-name="5">
-            <g isolation="isolate">
-            <polygon points="1721.5 558.77 1721.23 572.95 1565.55 548.3 1565.82 534.11 1721.5 558.77" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1565.82 534.11 1565.55 548.3 1507.78 679.8 1508.06 665.62 1565.82 534.11" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1663.73 690.28 1663.46 704.46 1507.78 679.8 1508.06 665.62 1663.73 690.28" fill="#19a6a6" strokeWidth="0px"/>
-            <polygon points="1721.5 558.77 1721.23 572.95 1663.46 704.46 1663.73 690.28 1721.5 558.77" fill="#1aa7a7" strokeWidth="0px"/>
-            <polygon points="1721.5 558.77 1663.73 690.28 1508.06 665.62 1565.82 534.11 1721.5 558.77" fill="#22dcdc" strokeWidth="0px"/>
-            <polygon points="1664.33 690.86 1664.05 705.04 1506.79 680.13 1507.06 665.95 1664.33 690.86" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1722.49 558.45 1722.22 572.63 1664.05 705.04 1664.33 690.86 1722.49 558.45" fill="#291455" strokeWidth="0px"/>
-            <path d="M1565.23,533.54l157.27,24.91-58.17,132.41-157.27-24.91,58.17-132.41ZM1663.73,690.28l57.77-131.5-155.68-24.66-57.77,131.5,155.68,24.66" fill="#361b70" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_6" data-name="6">
-            <g isolation="isolate">
-            <polygon points="1870.94 583.07 1870.67 597.25 1722.76 573.82 1723.03 559.64 1870.94 583.07" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1723.03 559.64 1722.76 573.82 1665.33 704.56 1665.6 690.38 1723.03 559.64" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1813.51 713.8 1813.24 727.98 1665.33 704.56 1665.6 690.38 1813.51 713.8" fill="#9f51a1" strokeWidth="0px"/>
-            <polygon points="1870.94 583.07 1870.67 597.25 1813.24 727.98 1813.51 713.8 1870.94 583.07" fill="#a051a2" strokeWidth="0px"/>
-            <path d="M1722.44,559.06l149.5,23.68-57.83,131.64-149.5-23.68,57.83-131.64ZM1813.51,713.8l57.43-130.74-147.91-23.43-57.43,130.74,147.91,23.43" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="1870.94 583.07 1813.51 713.8 1665.6 690.38 1723.03 559.64 1870.94 583.07" fill="#d36bd6" strokeWidth="0px"/>
-            <polygon points="1814.11 714.38 1813.83 728.56 1664.33 704.88 1664.61 690.7 1814.11 714.38" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1871.94 582.74 1871.66 596.92 1813.83 728.56 1814.11 714.38 1871.94 582.74" fill="#291455" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="_7" data-name="7">
-            <g isolation="isolate">
-            <polygon points="2116.66 583.8 2116.39 597.98 1999.55 579.47 1999.82 565.29 2116.66 583.8" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1999.82 565.29 1999.55 579.47 1984.09 614.66 1984.36 600.48 1999.82 565.29" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1984.36 600.48 1984.09 614.66 1872 596.91 1872.28 582.73 1984.36 600.48" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1872.28 582.73 1872 596.91 1814.4 728.03 1814.68 713.85 1872.28 582.73" fill="#291455" strokeWidth="0px"/>
-            <polygon points="2043.6 750.11 2043.33 764.29 1814.4 728.04 1814.68 713.85 2043.6 750.11" fill="#a6a600" strokeWidth="0px"/>
-            <polygon points="2116.66 583.8 2116.39 597.98 2043.33 764.29 2043.6 750.11 2116.66 583.8" fill="#a7a700" strokeWidth="0px"/>
-            <path d="M1999.23,564.72l118.42,18.76-73.45,167.21-230.5-36.51,58-132.03,112.08,17.75,15.46-35.19ZM2043.6,750.11l73.06-166.31-116.84-18.51-15.46,35.19-112.08-17.75-57.6,131.12,228.92,36.26" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="2116.66 583.8 2043.6 750.11 1814.68 713.85 1872.28 582.73 1984.36 600.48 1999.82 565.29 2116.66 583.8" fill="#dcdc00" strokeWidth="0px"/>
-            <polygon points="2044.2 750.69 2043.92 764.87 1813.42 728.36 1813.7 714.18 2044.2 750.69" fill="#281454" strokeWidth="0px"/>
-            <polygon points="2117.65 583.48 2117.38 597.66 2043.92 764.87 2044.2 750.69 2117.65 583.48" fill="#291455" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="pi1">
-            <g isolation="isolate">
-            <polygon points="907.96 384.08 907.64 400.7 899.3 399.38 899.62 382.76 907.96 384.08" fill="#281454" strokeWidth="0px"/>
-            <polygon points="899.62 382.76 899.3 399.38 897.2 404.17 897.52 387.55 899.62 382.76" fill="#291455" strokeWidth="0px"/>
-            <polygon points="905.86 388.87 905.54 405.49 897.2 404.17 897.52 387.55 905.86 388.87" fill="#421d5e" strokeWidth="0px"/>
-            <polygon points="907.96 384.08 907.64 400.7 905.54 405.49 905.86 388.87 907.96 384.08" fill="#431e5f" strokeWidth="0px"/>
-            <path d="M899.03,382.18l9.93,1.57-2.5,5.69-9.93-1.57,2.5-5.69ZM905.86,388.87l2.1-4.79-8.34-1.32-2.1,4.79,8.34,1.32" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="907.96 384.08 905.86 388.87 897.52 387.55 899.62 382.76 907.96 384.08" fill="#58277d" strokeWidth="0px"/>
-            <polygon points="906.46 389.45 906.14 406.06 896.21 404.49 896.53 387.87 906.46 389.45" fill="#281454" strokeWidth="0px"/>
-            <polygon points="908.96 383.75 908.64 400.37 906.14 406.06 906.46 389.45 908.96 383.75" fill="#291455" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="pi3">
-            <g isolation="isolate">
-            <polygon points="1032.99 406.02 1032.67 422.64 1024.32 421.31 1024.64 404.7 1032.99 406.02" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1024.64 404.7 1024.32 421.31 1022.22 426.1 1022.54 409.48 1024.64 404.7" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1030.89 410.81 1030.57 427.42 1022.22 426.1 1022.54 409.48 1030.89 410.81" fill="#421d5e" strokeWidth="0px"/>
-            <polygon points="1032.99 406.02 1032.67 422.64 1030.57 427.42 1030.89 410.81 1032.99 406.02" fill="#431e5f" strokeWidth="0px"/>
-            <polygon points="1032.99 406.02 1030.89 410.81 1022.54 409.48 1024.64 404.7 1032.99 406.02" fill="#58277d" strokeWidth="0px"/>
-            <polygon points="1031.48 411.39 1031.15 428 1021.23 426.43 1021.55 409.81 1031.48 411.39" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1033.98 405.69 1033.66 422.3 1031.15 428 1031.48 411.39 1033.98 405.69" fill="#291455" strokeWidth="0px"/>
-            <path d="M1024.06,404.12l9.92,1.57-2.5,5.7-9.92-1.57,2.5-5.7ZM1030.89,410.81l2.1-4.79-8.35-1.32-2.1,4.79,8.35,1.32" fill="#361b70" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="pi11">
-            <g isolation="isolate">
-            <polygon points="1824.62 529.01 1824.3 545.63 1815.95 544.31 1816.27 527.69 1824.62 529.01" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1816.27 527.69 1815.95 544.31 1813.85 549.1 1814.17 532.48 1816.27 527.69" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1822.51 533.8 1822.19 550.42 1813.85 549.1 1814.17 532.48 1822.51 533.8" fill="#421d5e" strokeWidth="0px"/>
-            <polygon points="1824.62 529.01 1824.3 545.63 1822.19 550.42 1822.51 533.8 1824.62 529.01" fill="#431e5f" strokeWidth="0px"/>
-            <path d="M1815.69,527.12l9.92,1.57-2.5,5.69-9.92-1.57,2.5-5.69ZM1822.51,533.8l2.1-4.79-8.34-1.32-2.1,4.79,8.34,1.32" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="1824.62 529.01 1822.51 533.8 1814.17 532.48 1816.27 527.69 1824.62 529.01" fill="#58277d" strokeWidth="0px"/>
-            <polygon points="1823.11 534.38 1822.79 551 1812.86 549.43 1813.19 532.81 1823.11 534.38" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1825.61 528.69 1825.29 545.3 1822.79 551 1823.11 534.38 1825.61 528.69" fill="#291455" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="pi12">
-            <g isolation="isolate">
-            <polygon points="1955.96 551.14 1955.64 567.75 1947.3 566.43 1947.62 549.81 1955.96 551.14" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1947.62 549.81 1947.3 566.43 1945.19 571.22 1945.52 554.61 1947.62 549.81" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1953.86 555.93 1953.54 572.55 1945.19 571.22 1945.52 554.61 1953.86 555.93" fill="#421d5e" strokeWidth="0px"/>
-            <polygon points="1955.96 551.14 1955.64 567.75 1953.54 572.55 1953.86 555.93 1955.96 551.14" fill="#431e5f" strokeWidth="0px"/>
-            <polygon points="1955.96 551.14 1953.86 555.93 1945.52 554.61 1947.62 549.82 1955.96 551.14" fill="#58277d" strokeWidth="0px"/>
-            <polygon points="1954.44 556.5 1954.12 573.12 1944.2 571.55 1944.52 554.93 1954.44 556.5" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1956.95 550.81 1956.62 567.43 1954.12 573.12 1954.44 556.5 1956.95 550.81" fill="#291455" strokeWidth="0px"/>
-            <path d="M1947.02,549.24l9.92,1.57-2.5,5.69-9.92-1.57,2.5-5.69ZM1953.86,555.93l2.11-4.79-8.34-1.32-2.11,4.79,8.34,1.32" fill="#361b70" strokeWidth="0px"/>
-            </g>
-        </g>
-        </svg>
-    </div>
-  )
-}
-
-export default NewAdminFloor1;
