--- src/api/new admin/floors/NewAdminFloor3.jsx
@@ -1,498 +1,0 @@
-import React from 'react'
-import { useContext, useState } from "react";
-
-export const NewAdminFloor3 = () => {
-
-    const [showInfoPanel, setShowInfoPanel] = useState(false);
-    const [roomInfo, setRoomInfo] = useState(null);
-    const [emoji, setEmoji] = useState("👷‍♂️");
-
-    const Floor3 = {
-    "3": {
-    "rooms": {
-        
-        }
-    }
-
-  }
-
-    const roomClicked = (key) =>{
-    setRoomInfo({ key, ...Floor1["1"].rooms[key] });
-    console.log(roomInfo);
-  } 
-
-    const roomOfs = -5;
-
-    const roomOnMouseEnter = (e, room, roomText) =>{
-    e.currentTarget.setAttribute('fill','#FFFFFF');
-    e.currentTarget.style.transform = `translate(${roomOfs}px, ${roomOfs}px)`; 
-    document.querySelector(roomText).style.transform = `translate(${roomOfs}px, ${roomOfs}px)`;
-    const roomStroke = document.querySelectorAll(room); 
-    console.log(roomStroke)
-    roomStroke.forEach(ro => {
-            ro.style.transform = `translate(${roomOfs}px, ${roomOfs}px)`;
-    });
-  }
-    const roomOnMouseLeave = (e, room, roomText, roomColor) =>{
-    e.currentTarget.setAttribute('fill', roomColor);
-    e.currentTarget.style.transform = 'translate(0px, 0px)'; 
-    document.querySelector(roomText).style.transform = 'translate(0px, 0px)'; 
-    const roomStroke = document.querySelectorAll(room); 
-    roomStroke.forEach(ro => {
-            ro.style.transform = 'translate(0px, 0px)'; 
-    });
-  }
-  return (
-    <div className='-mt-24 lg:flex w-screen justify-center items-center'>
-        <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 2348.7 1016.57">
-        <defs>
-            <clipPath id="clippath">
-            <path d="M568.28,460.01l-.27,13.73c-19,41.8-54.38,120.43-55.04,124.91l.27-13.73c.66-4.47,36.04-83.1,55.04-124.91Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-1">
-            <path d="M512.89,584.86l-.27,13.73s-.09-.01-.14-.01c-.07,0-.14.01-.21.03l.27-13.73c.06-.02.13-.03.21-.03.05,0,.09,0,.14.01Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-2">
-            <path d="M512.55,598.61c.09-4.58.18-9.16.27-13.73.05,0,.1,0,.16.01l-.27,13.73c-.06,0-.12,0-.16-.01Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-3">
-            <path d="M807.2,644.71c-257.53-40.79-289.88-45.6-294.22-46.06v-.06s.27-13.67.27-13.67c4.34.46,36.69,5.27,294.22,46.06,73.24,11.6,164.67,26.11,278.89,44.26,277.2,44.05,557.67,88.76,571.22,90.92l-.27,13.73c-13.55-2.16-294.02-46.87-571.22-90.92-114.22-18.15-205.64-32.66-278.89-44.26Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-4">
-            <path d="M512.33,598.62l.27-13.73c0,.43.17.49.35.54l-.27,13.73c-.18-.05-.36-.11-.35-.54Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-5">
-            <path d="M512.55,598.61s-.04,0-.06,0l.27-13.73s.04,0,.06,0c-.09,4.58-.18,9.16-.27,13.73Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-6">
-            <path d="M1658.31,766.87l-.27,13.73c-8.09-1.29-512.98-81.78-849.71-135.11-165.66-26.24-290.58-45.9-295.65-46.33l.27-13.73c5.07.43,129.99,20.09,295.65,46.33,336.73,53.33,841.62,133.82,849.71,135.11Z" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-7">
-            <polygon points="512.84 583.16 583.55 422.2 1729.48 603.7 1658.78 764.66 512.84 583.16" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-8">
-            <path d="M1352.87,582.49l15.79-35.94-425.69-67.42-15.68,35.68-358.52-57.01s-57.42,126.32-55.13,125.49c2.3-.84,1145.13,181.37,1145.13,181.37l54.82-124.8-360.73-57.36" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-9">
-            <polygon points="518.11 533.08 586.76 376.8 1730.48 557.95 1661.83 714.23 518.11 533.08" fill="none" strokeWidth="0px"/>
-            </clipPath>
-            <clipPath id="clippath-10">
-            <polyline points="1715.82 591.32 1661.83 714.23 518.11 533.08 572.31 409.69 930.23 466.54 944.6 433.83 1370.8 500.98 1356.57 533.38 1329.91 529.15 1329.84 529.32 1210.68 510.45 1216.84 496.45 1335.99 515.32 1337.7 511.41 967.24 452.74 965.52 456.64 1087.49 475.96 1081.2 490.29 959.23 470.97 959.16 471.13 1715.82 591.32" fill="none" strokeWidth="0px"/>
-            </clipPath>
-        </defs>
-        <g id="background">
-            <g id="back">
-            <path d="M1653.52,435.56l-355.67,123.25c-12.32,4.27-15.93,19.96-6.71,29.18l232.42,232.42c9.22,9.22,24.91,5.62,29.18-6.71l123.25-355.67c4.83-13.94-8.54-27.31-22.48-22.48Z" fill="rgba(41, 171, 226, .33)" strokeWidth="0px"/>
-            <path d="M1515.96,715.05h-808.64c-21.61,0-39.13-17.52-39.13-39.13v-200.35c0-21.61,17.52-39.13,39.13-39.13h738.2c60.51,0,109.57,49.05,109.57,109.57v129.91c0,21.61-17.52,39.13-39.13,39.13Z" fill="#ff7bac" opacity=".42" strokeWidth="0px"/>
-            <circle cx="743.33" cy="482.77" r="184.86" fill="rgba(0, 255, 0, .29)" strokeWidth="0px"/>
-            <path d="M1436.81,389.97l-49.54-7.45c-2.78-.42-5.57.67-7.33,2.86l-31.08,38.67c-1.78,2.22-2.23,5.22-1.18,7.86l18.34,46.1c1.04,2.61,3.39,4.47,6.16,4.89l49.54,7.45c2.78.42,5.57-.67,7.33-2.86l31.08-38.67c1.78-2.22,2.23-5.22,1.18-7.86l-18.34-46.1c-1.04-2.61-3.39-4.47-6.16-4.89Z" fill="rgba(145, 104, 176, .5)" opacity=".33" strokeWidth="0px"/>
-            <path d="M1002.55,602.84l-85.93-12.92c-2.78-.42-5.57.67-7.33,2.86l-53.96,67.14c-1.78,2.22-2.23,5.22-1.18,7.86l31.84,80.04c1.04,2.61,3.39,4.47,6.16,4.89l85.93,12.92c2.78.42,5.57-.67,7.33-2.86l53.96-67.14c1.78-2.22,2.23-5.22,1.18-7.86l-31.84-80.04c-1.04-2.61-3.39-4.47-6.16-4.89Z" fill="rgba(145, 104, 176, .5)" opacity=".33" strokeWidth="0px"/>
-            <path d="M1185.46,617.59l-34.69-5.21c-2.78-.42-5.57.67-7.33,2.86l-21.73,27.04c-1.78,2.22-2.23,5.22-1.18,7.86l12.82,32.24c1.04,2.61,3.39,4.47,6.16,4.89l34.69,5.21c2.78.42,5.57-.67,7.33-2.86l21.73-27.04c1.78-2.22,2.23-5.22,1.18-7.86l-12.82-32.24c-1.04-2.61-3.39-4.47-6.16-4.89Z" fill="rgba(145, 104, 176, .5)" opacity=".33" strokeWidth="0px"/>
-            <polygon points="1314.15 539.44 1216.47 520.04 1145.1 589.51 1133.37 490.61 1045.24 444.21 1135.68 402.49 1152.57 304.33 1220.2 377.45 1318.77 363.19 1270.13 450.1 1314.15 539.44" fill="rgba(255, 255, 0, .35)" strokeWidth="0px"/>
-            </g>
-        </g>
-        <g id="old_floors" data-name="old floors">
-            <g id="_1" data-name="1" opacity=".4">
-            </g>
-            <g id="_2" data-name="2">
-            <g id="floor">
-                <g isolation="isolate">
-                <polygon points="962.51 501.54 962.24 515.27 953.55 535.05 953.82 521.32 962.51 501.54" fill="#767676" strokeWidth="0px"/>
-                <polygon points="1333.19 560.25 1332.93 573.98 962.24 515.27 962.51 501.54 1333.19 560.25" fill="#747474" strokeWidth="0px"/>
-                <polygon points="963.24 502.25 962.97 515.99 961.41 519.55 961.67 505.82 963.24 502.25" fill="#212121" strokeWidth="0px"/>
-                <polygon points="961.67 505.82 961.41 519.55 960.43 519.4 960.7 505.67 961.67 505.82" fill="#212121" strokeWidth="0px"/>
-                <polygon points="942.49 481.34 942.22 495.07 926.54 530.76 926.81 517.02 942.49 481.34" fill="#212121" strokeWidth="0px"/>
-                <polygon points="926.81 517.02 926.54 530.76 568.01 473.74 568.28 460.01 926.81 517.02" fill="#212121" strokeWidth="0px"/>
-                <polygon points="961.43 506.38 961.16 520.11 954.62 535 954.89 521.26 961.43 506.38" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1331.97 560.66 1331.7 574.39 962.97 515.99 963.24 502.25 1331.97 560.66" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1083.22 525.67 1082.96 539.4 961.16 520.11 961.43 506.38 1083.22 525.67" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1083.22 525.67 1082.96 539.4 1076.42 554.29 1076.69 540.55 1083.22 525.67" fill="#646464" strokeWidth="0px"/>
-                <polygon points="1084.45 525.26 1084.18 538.99 1077.52 554.17 1077.78 540.43 1084.45 525.26" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1083.22 525.67 1076.69 540.55 954.89 521.26 961.43 506.38 1083.22 525.67" fill="#848484" strokeWidth="0px"/>
-                <polygon points="1076.69 540.55 1076.42 554.29 954.62 535 954.89 521.26 1076.69 540.55" fill="#636363" strokeWidth="0px"/>
-                <polygon points="1077.78 540.43 1077.52 554.17 1077.27 554.73 1077.54 540.99 1077.78 540.43" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1077.54 540.99 1077.27 554.73 1077.15 555 1077.42 541.27 1077.54 540.99" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1077.42 541.27 1077.15 555 953.4 535.4 953.66 521.67 1077.42 541.27" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1367.19 548.61 1366.93 562.34 942.22 495.07 942.49 481.34 1367.19 548.61" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1208.74 545.36 1208.48 559.1 1201.94 573.98 1202.2 560.25 1208.74 545.36" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1201.24 560.06 1200.97 573.8 1200.73 574.36 1200.99 560.62 1201.24 560.06" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1077.54 540.99 1077.78 540.43 1201.24 560.06 1200.99 560.62 1077.54 540.99" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="1200.99 560.62 1200.73 574.36 1077.27 554.73 1077.54 540.99 1200.99 560.62" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1333.19 560.25 1332.93 573.98 1331.19 577.93 1331.46 564.2 1333.19 560.25" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1331.46 564.2 1331.19 577.93 1330.22 577.77 1330.48 564.04 1331.46 564.2" fill="#212121" strokeWidth="0px"/>
-                <polygon points="960.7 505.67 962.51 501.54 1333.19 560.25 1331.46 564.2 1330.48 564.04 1331.97 560.66 963.24 502.25 961.67 505.82 960.7 505.67" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="1330.54 564.65 1330.27 578.39 1208.48 559.1 1208.74 545.36 1330.54 564.65" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1324 579.54 1323.73 593.28 1201.94 573.99 1202.2 560.25 1324 579.54" fill="#636363" strokeWidth="0px"/>
-                <polygon points="1256.62 569.47 1256.35 583.2 1200.71 574.39 1200.98 560.66 1256.62 569.47" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1330.54 564.66 1330.27 578.39 1323.73 593.28 1324 579.54 1330.54 564.66" fill="#646464" strokeWidth="0px"/>
-                <polygon points="1330.54 564.66 1324 579.54 1202.2 560.25 1208.74 545.36 1330.54 564.66" fill="#848484" strokeWidth="0px"/>
-                <polygon points="1324.42 580.21 1324.16 593.94 1256.35 583.2 1256.62 569.47 1324.42 580.21" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1323.91 580.17 1323.64 593.9 1256.35 583.2 1256.62 569.47 1323.91 580.17" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1324.42 580.21 1324.16 593.94 1323.64 593.9 1323.91 580.17 1324.42 580.21" fill="#232323" strokeWidth="0px"/>
-                <polygon points="1324.73 580.26 1324.46 593.99 1324.16 593.94 1324.42 580.21 1324.73 580.26" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1331.76 564.24 1331.5 577.98 1324.46 593.99 1324.73 580.26 1331.76 564.24" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1367.19 548.61 1366.93 562.34 1351.14 598.28 1351.4 584.55 1367.19 548.61" fill="#767676" strokeWidth="0px"/>
-                <polygon points="1368.41 548.2 1368.15 561.93 1352.36 597.87 1352.63 584.14 1368.41 548.2" fill="#212121" strokeWidth="0px"/>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath)">
-                    <g isolation="isolate">
-                        <path d="M568.28,460.01l-.27,13.73c-15.13,33.28-40.63,89.9-50.74,113.86l.27-13.73c10.11-23.96,35.61-80.58,50.74-113.86" fill="#212121" strokeWidth="0px"/>
-                        <path d="M517.54,573.87l-.27,13.73c-2.28,5.42-3.78,9.17-4.19,10.6l.27-13.73c.41-1.43,1.91-5.18,4.19-10.6" fill="#202020" strokeWidth="0px"/>
-                        <path d="M513.35,584.47l-.27,13.73c-.05.19-.09.34-.11.45l.27-13.73c.02-.11.05-.26.11-.45" fill="#1f1f1f" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <polygon points="1712.15 641.91 1711.88 655.65 1351.14 598.28 1351.4 584.55 1712.15 641.91" fill="#212121" strokeWidth="0px"/>
-                <path d="M1351.4,584.55l360.75,57.37-54.57,124.23c-13.55-2.16-294.02-46.87-571.22-90.92-114.22-18.15-205.64-32.66-278.89-44.26-257.53-40.79-289.88-45.6-294.22-46.06.66-4.48,36.04-83.1,55.04-124.9l358.04,56.94,16.16-35.61,424.7,67.27-15.79,35.94ZM1323.91,580.17l7.85-15.93,1.43-3.99-370.68-58.71-8.69,19.78,123.6,19.95,123.58,19.36,55.63,8.85,67.29,10.7" fill="#9b9b9b" strokeWidth="0px"/>
-                <polygon points="512.99 584.88 512.72 598.61 512.63 598.59 512.89 584.86 512.99 584.88" fill="#202020" strokeWidth="0px"/>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath-1)">
-                    <g isolation="isolate">
-                        <polyline points="512.89 584.86 512.63 598.59 512.63 598.59 512.89 584.86 512.89 584.86" fill="#1e1e1e" strokeWidth="0px"/>
-                        <path d="M512.89,584.86l-.27,13.73s-.03,0-.05,0l.27-13.73s.03,0,.05,0" fill="#1f1f1f" strokeWidth="0px"/>
-                        <path d="M512.84,584.85l-.27,13.73s-.03,0-.05,0l.27-13.73s.03,0,.05,0" fill="#202020" strokeWidth="0px"/>
-                        <path d="M512.8,584.85l-.27,13.73s-.03,0-.04,0c0,0,0,0-.01,0l.27-13.73s0,0,.01,0c.01,0,.03,0,.04,0" fill="#212121" strokeWidth="0px"/>
-                        <path d="M512.74,584.85l-.27,13.73s-.03,0-.05,0l.27-13.73s.03,0,.05,0" fill="#222" strokeWidth="0px"/>
-                        <path d="M512.7,584.85l-.27,13.73s-.04,0-.06,0l.27-13.73s.04,0,.06,0" fill="#232323" strokeWidth="0px"/>
-                        <path d="M512.64,584.86l-.27,13.73s-.06.01-.09.02l.27-13.73s.06-.02.09-.02" fill="#242424" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath-2)">
-                    <g isolation="isolate">
-                        <path d="M512.97,584.89l-.27,13.73c-.09,0-.16-.01-.22-.02l.27-13.73c.06,0,.13,0,.22.02" fill="#6f6f6f" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <path d="M512.76,584.85s.09,0,.14.01l.09.02h-.01c-.09,0-.16,0-.22,0,0,0,0-.01,0-.02Z" fill="#9b9b9b" strokeWidth="0px"/>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath-3)">
-                    <g isolation="isolate">
-                        <path d="M1657.58,766.15l-.27,13.73c-13.55-2.16-294.02-46.87-571.22-90.92-114.22-18.15-205.64-32.66-278.89-44.26-58.28-9.23-105.03-16.62-142.53-22.53l.27-13.73c37.5,5.91,84.25,13.3,142.53,22.53,73.24,11.6,164.67,26.11,278.89,44.26,277.2,44.05,557.67,88.76,571.22,90.92" fill="#686868" strokeWidth="0px"/>
-                        <path d="M664.94,608.44l-.27,13.73c-128.21-20.21-148.34-23.17-151.69-23.53l.27-13.73c3.36.36,23.48,3.32,151.69,23.53" fill="#6b6b6b" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <polygon points="1712.15 641.91 1711.89 655.65 1657.31 779.88 1657.58 766.15 1712.15 641.91" fill="#767676" strokeWidth="0px"/>
-                <path d="M1352.63,584.14l360.75,57.37-55.07,125.36c-8.09-1.29-512.98-81.78-849.71-135.11-165.66-26.24-290.58-45.9-295.65-46.33-1-.29-1.74-.6,26.32-63.36,13.82-30.93,28.02-62.18,28.17-62.49l358.65,56.73,15.67-35.68,426.66,67.58-15.79,35.94ZM1657.58,766.15l54.57-124.23-360.75-57.37,15.79-35.94-424.71-67.27-15.68,35.68-358.53-57.02c-19,41.8-54.38,120.43-55.04,124.91,4.34.46,36.69,5.27,294.22,46.06,73.25,11.6,164.67,26.11,278.89,44.26,277.21,44.05,557.68,88.76,571.22,90.92M512.98,584.89h.01s-.09-.03-.09-.03c-.05,0-.09-.01-.14-.01-.07,0-.14.01-.21.03.03,0,.08-.01.2,0,.06,0,.13,0,.22.02" fill="#2c2c2c" strokeWidth="0px"/>
-                <path d="M961.67,505.82l122.78,19.44-6.66,15.17-.25.56-.12.28-123.75-19.6.15-.35,6.88-15.65.98.15ZM1076.68,540.55l6.54-14.89-121.8-19.29-6.54,14.88,121.8,19.29" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="1323.91 580.17 1256.62 569.47 1324.42 580.21 1323.91 580.17" fill="#2c2c2c" strokeWidth="0px"/>
-                <path d="M1331.46,564.2l.3.05-7.04,16.01-.3-.05-67.8-10.74-55.64-8.81.26-.59,6.77-15.42,122.47,19.39.98.15ZM1324,579.54l6.54-14.89-121.8-19.29-6.54,14.89,121.8,19.29" fill="#2c2c2c" strokeWidth="0px"/>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath-4)">
-                    <g isolation="isolate">
-                        <path d="M512.94,585.43l-.27,13.73s-.01,0-.02,0l.27-13.73s.01,0,.02,0" fill="#1d1d1d" strokeWidth="0px"/>
-                        <path d="M512.93,585.42l-.27,13.73s-.06-.02-.09-.03l.27-13.73s.06.02.09.03" fill="#1c1c1c" strokeWidth="0px"/>
-                        <path d="M512.84,585.39l-.27,13.73c-.14-.05-.25-.16-.24-.51l.27-13.73c0,.35.1.45.24.51" fill="#1b1b1b" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath-5)">
-                    <g isolation="isolate">
-                        <path d="M512.97,584.89l-.27,13.73c-.09,0-.16-.01-.22-.02l.27-13.73c.06,0,.13,0,.22.02" fill="#6f6f6f" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <g isolation="isolate">
-                    <g clip-path="url(#clippath-6)">
-                    <g isolation="isolate">
-                        <path d="M1658.31,766.87l-.27,13.73c-8.09-1.29-512.98-81.78-849.71-135.11-55.02-8.71-105.54-16.7-148.67-23.5l.27-13.73c43.13,6.8,93.66,14.79,148.67,23.5,336.73,53.33,841.62,133.82,849.71,135.11" fill="#1e1e1e" strokeWidth="0px"/>
-                        <path d="M659.92,608.26l-.27,13.73c-85.11-13.42-141.45-22.21-146.73-22.8l.27-13.73c5.29.6,61.62,9.38,146.73,22.8" fill="#1f1f1f" strokeWidth="0px"/>
-                        <path d="M513.19,585.45l-.27,13.73c-.1-.01-.18-.02-.25-.03l.27-13.73c.06,0,.15.01.25.03" fill="#202020" strokeWidth="0px"/>
-                    </g>
-                    </g>
-                </g>
-                <polygon points="1713.37 641.51 1713.11 655.24 1658.04 780.6 1658.31 766.87 1713.37 641.51" fill="#212121" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g opacity=".17">
-                <g isolation="isolate">
-                <g clip-path="url(#clippath-7)">
-                    <g clip-path="url(#clippath-8)">
-                    <path d="M560.62,576.81l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M616.14,585.6l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M671.66,594.4l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M727.17,603.19l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M782.7,611.98l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M838.21,620.78l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M893.73,629.57l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M949.25,638.36l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1004.76,647.15l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1060.28,655.95l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1115.8,664.74l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1171.32,673.54l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1226.84,682.33l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1282.35,691.12l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1337.87,699.91l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1393.39,708.71l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1448.91,717.5l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1504.42,726.29l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1559.95,735.09l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M569.82,555.88l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M625.33,564.67l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M680.86,573.47l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M736.37,582.26l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M791.89,591.05l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M847.41,599.84l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M902.93,608.64l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M958.44,617.43l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1013.96,626.22l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1069.48,635.02l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1124.99,643.81l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1180.52,652.6l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1236.03,661.4l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1291.54,670.19l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1347.07,678.98l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1402.58,687.78l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1458.1,696.57l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1513.62,705.36l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1569.14,714.16l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M579.01,534.95l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M634.53,543.74l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M690.05,552.54l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M745.56,561.33l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M801.08,570.12l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M856.6,578.92l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M912.12,587.71l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M967.64,596.5l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1023.15,605.3l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1078.67,614.09l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1134.19,622.88l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1189.71,631.68l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1245.22,640.47l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1300.74,649.26l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1356.26,658.06l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1411.77,666.85l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1467.3,675.64l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1522.81,684.43l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1578.33,693.23l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M588.21,514.02l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M643.72,522.81l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M699.24,531.61l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M754.76,540.4l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M810.28,549.19l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M865.79,557.99l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M921.32,566.78l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M976.83,575.57l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1032.34,584.36l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1087.87,593.16l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1143.38,601.95l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1198.9,610.75l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1254.42,619.54l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1309.93,628.33l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1365.45,637.12l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1420.97,645.92l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1476.49,654.71l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1532.01,663.5l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1587.53,672.3l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M597.4,493.09l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M652.91,501.89l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M708.44,510.68l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M763.95,519.47l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M819.47,528.27l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M874.99,537.06l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M930.51,545.85l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M986.02,554.65l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1041.54,563.44l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1097.06,572.23l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1152.57,581.02l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1208.1,589.82l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1263.61,598.61l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1319.13,607.4l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1374.65,616.2l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1430.16,624.99l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1485.68,633.78l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1541.2,642.58l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1596.72,651.37l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M939.7,524.93l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M995.22,533.72l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1050.73,542.51l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1106.25,551.31l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1161.77,560.1l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1217.29,568.89l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1272.8,577.68l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M948.9,503.99l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1004.41,512.79l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1059.93,521.58l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1115.45,530.37l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1170.96,539.17l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1226.48,547.96l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1282,556.75l8.96-20.39,54.58,8.64-8.96,20.39-54.58-8.64M1347.34,543.17l-.93-.15-.64,1.45-54.58-8.64.64-1.45-.93-.15-.64,1.45-54.58-8.64.64-1.45-.94-.15-.64,1.45-54.58-8.64.64-1.45-.93-.15-.64,1.45-54.58-8.64.64-1.45-.94-.15-.64,1.45-54.58-8.64.64-1.45-.93-.15-.64,1.45-54.58-8.64.64-1.45-.93-.15-.64,1.45-54.58-8.64.64-1.45-.94-.15-.64,1.45-14.82-2.35-.24.54,14.82,2.35-8.96,20.39-14.82-2.35-.24.54,14.82,2.35-8.96,20.39-54.58-8.64,3.36-7.65-.93-.15-3.36,7.65-54.58-8.64,3.37-7.68-.94-.15-3.37,7.68-54.58-8.64,3.39-7.71-.93-.15-3.39,7.71-54.58-8.64,3.4-7.75-.94-.15-3.4,7.75-54.58-8.64,3.42-7.78-.93-.15-3.42,7.78-54.58-8.64,3.43-7.81-.93-.15-3.43,7.81-40.44-6.4c-.08.17-.16.36-.24.53l40.44,6.41-8.96,20.39-40.69-6.44c-.08.18-.16.36-.24.53l40.69,6.45-8.96,20.39-40.9-6.48c-.08.18-.16.36-.24.54l40.9,6.48-8.96,20.39-41.06-6.5c-.08.18-.16.36-.24.54l41.06,6.5-8.96,20.39-41.13-6.51c-.08.18-.16.36-.24.54l41.13,6.51-8.96,20.39-41.03-6.5c-.08.18-.15.36-.23.54l41.02,6.5-5.35,12.18c.31.05.62.1.93.15l5.35-12.17,54.58,8.64-5.3,12.07c.31.05.63.1.93.15l5.3-12.07,54.58,8.64-5.28,12.02c.32.05.62.1.94.15l5.28-12.01,54.58,8.64-5.27,11.99c.33.05.6.1.93.15l5.27-11.99,54.58,8.64-5.26,11.98c.33.05.61.1.94.15l5.26-11.98,54.58,8.64-5.26,11.97c.32.05.62.1.93.15l5.26-11.97,54.58,8.64-5.26,11.98c.33.05.61.1.94.15l5.26-11.98,54.58,8.64-5.27,11.99c.31.05.63.1.93.15l5.27-11.99,54.58,8.64-5.28,12.01c.31.05.62.1.93.15l5.28-12.01,54.58,8.64-5.29,12.04c.31.05.63.1.94.15l5.29-12.04,54.58,8.64-5.3,12.06c.31.05.62.1.93.15l5.3-12.06,54.58,8.64-5.31,12.09c.31.05.63.1.94.15l5.31-12.09,54.58,8.64-5.33,12.13c.31.05.63.1.93.15l5.33-12.13,54.58,8.64-5.34,12.16c.3.05.64.1.93.15l5.34-12.16,54.58,8.64-5.36,12.2c.32.05.62.1.94.15l5.36-12.2,54.58,8.64-5.38,12.24c.33.05.6.1.93.15l5.38-12.24,54.58,8.64-5.4,12.28c.31.05.63.1.94.15l5.4-12.29,54.58,8.64-5.42,12.33c.33.05.61.1.93.15l5.42-12.33,54.58,8.64-5.44,12.38c.32.05.62.1.94.15l5.44-12.38,54.58,8.64-5.46,12.43c.31.05.63.1.93.15l5.46-12.43,49.03,7.77.24-.54-49.03-7.77,8.96-20.39,49.03,7.77.24-.54-49.03-7.77,8.96-20.39,49.03,7.77.24-.54-49.03-7.77,8.96-20.39,49.03,7.77.24-.54-49.03-7.77,8.96-20.39,49.03,7.77.24-.54-49.03-7.77,8.96-20.39,49.03,7.77.24-.54-49.03-7.77,3.15-7.18-.93-.15-3.15,7.18-54.58-8.64,3.17-7.21-.94-.15-3.17,7.21-54.58-8.64,3.18-7.25-.93-.15-3.18,7.25-54.58-8.64,3.2-7.28-.94-.15-3.2,7.28-54.58-8.64,3.21-7.31-.93-.15-3.21,7.31-54.58-8.64,3.23-7.35-.94-.15-3.23,7.35-54.58-8.64,8.96-20.39,21.32,3.38.24-.54-21.32-3.38,8.96-20.39,21.32,3.38.24-.54-21.32-3.38.64-1.45" fill="#fff" strokeWidth="0px"/>
-                    </g>
-                </g>
-                </g>
-            </g>
-            <g id="_1-2" data-name="1">
-                <g isolation="isolate">
-                <polygon points="661.67 470.29 661.35 487.17 563.7 471.7 564.03 454.83 661.67 470.29" fill="#212121" strokeWidth="0px"/>
-                <polygon points="661.67 470.29 661.35 487.17 649.8 512.95 650.13 496.08 661.67 470.29" fill="#7c7c7c" strokeWidth="0px"/>
-                <polygon points="662.85 469.9 662.53 486.78 650.98 512.56 651.31 495.69 662.85 469.9" fill="#222" strokeWidth="0px"/>
-                <polygon points="670.09 499.24 669.76 516.11 649.8 512.95 650.13 496.08 670.09 499.24" fill="#212121" strokeWidth="0px"/>
-                <polygon points="564.03 454.83 563.7 471.7 513.71 585.51 514.03 568.64 564.03 454.83" fill="#212121" strokeWidth="0px"/>
-                <polygon points="661.67 470.29 650.13 496.08 670.09 499.24 631.43 587.24 514.03 568.64 564.03 454.83 661.67 470.29" fill="#a3a3a3" strokeWidth="0px"/>
-                <polygon points="631.43 587.24 631.11 604.11 513.71 585.51 514.03 568.64 631.43 587.24" fill="#7a7a7a" strokeWidth="0px"/>
-                <polygon points="670.09 499.24 669.76 516.11 631.11 604.11 631.43 587.24 670.09 499.24" fill="#7b7b7b" strokeWidth="0px"/>
-                <path d="M651.31,495.69l19.95,3.16-39.13,89.08-119.27-18.89,50.47-114.89,99.53,15.76-11.54,25.79ZM631.43,587.24l38.65-87.99-19.96-3.16,11.54-25.79-97.64-15.47-50,113.81,117.4,18.59" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="671.26 498.85 670.93 515.72 631.8 604.8 632.13 587.92 671.26 498.85" fill="#212121" strokeWidth="0px"/>
-                <polygon points="632.13 587.92 631.8 604.8 512.53 585.91 512.85 569.03 632.13 587.92" fill="#212121" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_2-2" data-name="2">
-                <g isolation="isolate">
-                <polygon points="734.83 509.09 734.51 525.96 671.7 516.01 672.02 499.14 734.83 509.09" fill="#212121" strokeWidth="0px"/>
-                <polygon points="672.02 499.14 671.7 516.01 632.9 604.33 633.23 587.46 672.02 499.14" fill="#212121" strokeWidth="0px"/>
-                <polygon points="696.04 597.4 695.71 614.28 632.9 604.33 633.23 587.46 696.04 597.4" fill="#6d6d6d" strokeWidth="0px"/>
-                <polygon points="734.83 509.09 734.51 525.96 695.71 614.28 696.04 597.4 734.83 509.09" fill="#6e6e6e" strokeWidth="0px"/>
-                <polygon points="734.83 509.09 696.04 597.4 633.23 587.46 672.02 499.14 734.83 509.09" fill="#919191" strokeWidth="0px"/>
-                <polygon points="696.74 598.09 696.41 614.96 631.73 604.72 632.06 587.84 696.74 598.09" fill="#212121" strokeWidth="0px"/>
-                <polygon points="736 508.7 735.68 525.57 696.41 614.96 696.74 598.09 736 508.7" fill="#212121" strokeWidth="0px"/>
-                <path d="M671.33,498.45l64.68,10.24-39.27,89.39-64.68-10.24,39.27-89.39ZM696.04,597.4l38.79-88.31-62.81-9.95-38.79,88.31,62.81,9.95" fill="#2c2c2c" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_3" data-name="3">
-                <g isolation="isolate">
-                <polygon points="812.39 521.5 812.06 538.38 735.48 526.08 735.8 509.21 812.39 521.5" fill="#212121" strokeWidth="0px"/>
-                <polygon points="735.8 509.21 735.48 526.08 696.75 614.23 697.08 597.36 735.8 509.21" fill="#212121" strokeWidth="0px"/>
-                <polygon points="773.73 609.5 773.41 626.37 696.75 614.23 697.08 597.36 773.73 609.5" fill="#8a8a8a" strokeWidth="0px"/>
-                <polygon points="812.39 521.5 812.06 538.38 773.41 626.37 773.73 609.5 812.39 521.5" fill="#8b8b8b" strokeWidth="0px"/>
-                <path d="M735.1,508.52l78.47,12.6-39.13,89.08-78.54-12.44,39.2-89.24ZM773.73,609.5l38.66-88-76.59-12.3-38.72,88.15,76.66,12.14" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="735.8 509.21 812.39 521.5 773.73 609.5 697.08 597.36 735.8 509.21" fill="#b7b7b7" strokeWidth="0px"/>
-                <polygon points="774.44 610.19 774.11 627.06 695.57 614.62 695.9 597.75 774.44 610.19" fill="#212121" strokeWidth="0px"/>
-                <polygon points="813.57 521.12 813.24 537.99 774.11 627.06 774.44 610.19 813.57 521.12" fill="#212121" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_4" data-name="4">
-                <g isolation="isolate">
-                <polygon points="935.47 540.64 935.15 557.51 813.1 538.01 813.43 521.14 935.47 540.64" fill="#212121" strokeWidth="0px"/>
-                <polygon points="813.43 521.14 813.1 538.01 774.45 626.01 774.78 609.13 813.43 521.14" fill="#212121" strokeWidth="0px"/>
-                <polygon points="896.89 628.47 896.56 645.35 774.45 626.01 774.78 609.13 896.89 628.47" fill="#6a6a6a" strokeWidth="0px"/>
-                <polygon points="935.47 540.64 935.15 557.51 896.56 645.35 896.89 628.47 935.47 540.64" fill="#6b6b6b" strokeWidth="0px"/>
-                <polygon points="813.43 521.14 935.47 540.64 896.89 628.47 774.78 609.13 813.43 521.14" fill="#8d8d8d" strokeWidth="0px"/>
-                <polygon points="897.59 629.16 897.27 646.04 773.27 626.4 773.6 609.53 897.59 629.16" fill="#212121" strokeWidth="0px"/>
-                <polygon points="936.65 540.25 936.32 557.12 897.27 646.04 897.59 629.16 936.65 540.25" fill="#212121" strokeWidth="0px"/>
-                <path d="M812.73,520.45l123.92,19.8-39.06,88.92-123.99-19.64,39.13-89.08ZM896.89,628.47l38.59-87.84-122.04-19.5-38.65,87.99,122.11,19.34" fill="#2c2c2c" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_5" data-name="5">
-                <g isolation="isolate">
-                <polygon points="1056.23 559.73 1055.91 576.6 936.57 557.7 936.9 540.83 1056.23 559.73" fill="#212121" strokeWidth="0px"/>
-                <polygon points="936.9 540.83 936.57 557.7 897.98 645.55 898.31 628.67 936.9 540.83" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1017.64 647.57 1017.32 664.45 897.98 645.55 898.31 628.67 1017.64 647.57" fill="#606060" strokeWidth="0px"/>
-                <polygon points="1056.23 559.73 1055.91 576.6 1017.32 664.45 1017.64 647.57 1056.23 559.73" fill="#616161" strokeWidth="0px"/>
-                <polygon points="1056.23 559.73 1017.64 647.57 898.31 628.67 936.9 540.83 1056.23 559.73" fill="gray" strokeWidth="0px"/>
-                <polygon points="1018.35 648.26 1018.03 665.13 896.81 645.93 897.13 629.06 1018.35 648.26" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1057.41 559.34 1057.08 576.22 1018.03 665.13 1018.35 648.26 1057.41 559.34" fill="#212121" strokeWidth="0px"/>
-                <path d="M936.19,540.14l121.22,19.2-39.06,88.91-121.22-19.2,39.06-88.91ZM1017.64,647.57l38.59-87.84-119.34-18.9-38.59,87.84,119.34,18.9" fill="#2c2c2c" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_7" data-name="7">
-                <g isolation="isolate">
-                <polygon points="1330.03 603.78 1329.7 620.65 1193.19 599.03 1193.51 582.15 1330.03 603.78" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1193.51 582.15 1193.19 599.03 1154.67 686.7 1155 669.83 1193.51 582.15" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1291.51 691.45 1291.19 708.33 1154.67 686.7 1155 669.83 1291.51 691.45" fill="#6d6d6d" strokeWidth="0px"/>
-                <polygon points="1330.03 603.78 1329.7 620.65 1291.19 708.33 1291.51 691.45 1330.03 603.78" fill="#6e6e6e" strokeWidth="0px"/>
-                <polygon points="1330.03 603.78 1291.51 691.45 1155 669.83 1193.51 582.15 1330.03 603.78" fill="#919191" strokeWidth="0px"/>
-                <polygon points="1292.22 692.14 1291.89 709.02 1153.49 687.1 1153.82 670.22 1292.22 692.14" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1331.21 603.38 1330.88 620.26 1291.89 709.02 1292.22 692.14 1331.21 603.38" fill="#212121" strokeWidth="0px"/>
-                <path d="M1192.81,581.46l138.4,21.92-38.99,88.76-138.4-21.92,38.99-88.76ZM1291.51,691.45l38.52-87.68-136.52-21.62-38.52,87.68,136.52,21.62" fill="#2c2c2c" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_8" data-name="8">
-                <g isolation="isolate">
-                <polygon points="1580.97 643.41 1580.65 660.28 1331.61 620.84 1331.94 603.97 1580.97 643.41" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1331.94 603.97 1331.61 620.84 1293.09 708.52 1293.42 691.65 1331.94 603.97" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1542.46 731.1 1542.13 747.97 1293.09 708.52 1293.42 691.65 1542.46 731.1" fill="#5d5d5d" strokeWidth="0px"/>
-                <polygon points="1580.97 643.41 1580.65 660.28 1542.13 747.97 1542.46 731.1 1580.97 643.41" fill="#5e5e5e" strokeWidth="0px"/>
-                <path d="M1331.23,603.28l250.92,39.74-38.99,88.76-250.92-39.74,38.99-88.76ZM1542.46,731.1l38.52-87.68-249.04-39.44-38.52,87.68,249.04,39.44" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="1580.97 643.41 1542.46 731.1 1293.42 691.65 1331.94 603.97 1580.97 643.41" fill="#7c7c7c" strokeWidth="0px"/>
-                <polygon points="1543.16 731.78 1542.84 748.65 1291.92 708.91 1292.24 692.04 1543.16 731.78" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1582.15 643.02 1581.83 659.89 1542.84 748.65 1543.16 731.78 1582.15 643.02" fill="#212121" strokeWidth="0px"/>
-                </g>
-            </g>
-            <g id="_9" data-name="9">
-                <g isolation="isolate">
-                <polygon points="1707.45 635.88 1707.12 652.75 1609.68 637.32 1610 620.45 1707.45 635.88" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1610 620.45 1609.68 637.32 1598.41 662.98 1598.73 646.11 1610 620.45" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1598.73 646.11 1598.41 662.98 1579.84 660.04 1580.16 643.16 1598.73 646.11" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1580.16 643.16 1579.84 660.04 1541.39 747.56 1541.72 730.69 1580.16 643.16" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1707.45 635.88 1657.73 749.06 1541.72 730.69 1580.16 643.16 1598.73 646.11 1610.01 620.45 1707.45 635.88" fill="#909090" strokeWidth="0px"/>
-                <polygon points="1657.73 749.06 1657.4 765.93 1541.39 747.56 1541.72 730.69 1657.73 749.06" fill="#6c6c6c" strokeWidth="0px"/>
-                <polygon points="1707.45 635.88 1707.12 652.75 1657.4 765.93 1657.73 749.06 1707.45 635.88" fill="#6d6d6d" strokeWidth="0px"/>
-                <path d="M1609.3,619.76l99.33,15.73-50.19,114.26-117.9-18.67,38.92-88.6,18.57,2.94,11.27-25.66ZM1657.73,749.06l49.72-113.18-97.44-15.43-11.27,25.66-18.57-2.94-38.45,87.52,116.01,18.37" fill="#2c2c2c" strokeWidth="0px"/>
-                <polygon points="1658.43 749.75 1658.11 766.62 1540.21 747.95 1540.54 731.08 1658.43 749.75" fill="#212121" strokeWidth="0px"/>
-                <polygon points="1708.63 635.49 1708.3 652.36 1658.11 766.62 1658.43 749.75 1708.63 635.49" fill="#212121" strokeWidth="0px"/>
-                </g>
-            </g>
-            </g>
-        </g>
-        <g id="Layer_2" data-name="Layer 2">
-            <g id="floor-2" data-name="floor">
-            <g isolation="isolate">
-                <polygon points="962.01 457.29 961.74 471.47 953.59 490.03 953.86 475.85 962.01 457.29" fill="#767676" strokeWidth="0px"/>
-                <polygon points="961.27 461.35 960.99 475.53 960.02 475.38 960.3 461.2 961.27 461.35" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1333.44 516.12 1333.16 530.3 961.74 471.47 962.01 457.29 1333.44 516.12" fill="#747474" strokeWidth="0px"/>
-                <polygon points="962.74 458 962.46 472.18 960.99 475.53 961.27 461.35 962.74 458" fill="#291455" strokeWidth="0px"/>
-                <polygon points="925 471.09 924.73 485.27 567.54 429.13 567.81 414.95 925 471.09" fill="#281454" strokeWidth="0px"/>
-                <polygon points="940.1 439.09 939.82 453.27 925.62 485.6 925.89 471.42 940.1 439.09" fill="#291455" strokeWidth="0px"/>
-                <polygon points="953.86 475.85 953.59 490.03 925.62 485.6 925.89 471.42 953.86 475.85" fill="#747474" strokeWidth="0px"/>
-                <polygon points="960.3 461.2 960.02 475.38 953.49 490.26 953.76 476.08 960.3 461.2" fill="#291455" strokeWidth="0px"/>
-                <polygon points="961.02 461.91 960.75 476.09 954.7 489.86 954.97 475.68 961.02 461.91" fill="#291455" strokeWidth="0px"/>
-                <polygon points="954.73 476.23 954.45 490.41 953.48 490.26 953.76 476.08 954.73 476.23" fill="#281454" strokeWidth="0px"/>
-                <polygon points="954.73 476.24 954.45 490.42 924.41 486 924.68 471.82 954.73 476.24" fill="#291454" strokeWidth="0px"/>
-                <polygon points="1332.23 516.52 1331.95 530.7 962.46 472.18 962.74 458 1332.23 516.52" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1082.02 481.07 1081.75 495.25 960.75 476.09 961.02 461.91 1082.02 481.07" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1082.02 481.07 1075.97 494.84 954.97 475.68 961.02 461.91 1082.02 481.07" fill="#6f6f6f" strokeWidth="0px"/>
-                <polygon points="1082.02 481.07 1081.75 495.25 1075.7 509.02 1075.97 494.84 1082.02 481.07" fill="#545454" strokeWidth="0px"/>
-                <polygon points="1075.97 494.84 1075.7 509.02 954.7 489.86 954.97 475.68 1075.97 494.84" fill="#535353" strokeWidth="0px"/>
-                <polygon points="1083.23 480.67 1082.96 494.85 1076.57 509.39 1076.85 495.21 1083.23 480.67" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1365.33 506.09 1365.05 520.28 939.82 453.27 940.1 439.09 1365.33 506.09" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1330.51 520.43 1330.24 534.61 1212.07 515.89 1212.34 501.71 1330.51 520.43" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1331.72 520.03 1331.45 534.21 1330.48 534.05 1330.75 519.87 1331.72 520.03" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1330.51 520.43 1330.24 534.61 1324.33 548.06 1324.6 533.88 1330.51 520.43" fill="#545454" strokeWidth="0px"/>
-                <polygon points="1331.72 520.03 1331.45 534.21 1325.37 548.04 1325.64 533.86 1331.72 520.03" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1325.64 533.86 1325.37 548.04 1325.13 548.6 1325.4 534.42 1325.64 533.86" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1333.44 516.12 1333.16 530.3 1331.45 534.21 1331.72 520.03 1333.44 516.12" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1351.34 537.93 1351.07 552.11 1325.37 548.04 1325.64 533.86 1351.34 537.93" fill="#747474" strokeWidth="0px"/>
-                <polygon points="1365.33 506.09 1365.05 520.28 1351.07 552.11 1351.34 537.93 1365.33 506.09" fill="#767676" strokeWidth="0px"/>
-                <polygon points="1366.54 505.69 1352.07 538.64 1325.4 534.42 1325.64 533.86 1351.34 537.93 1365.33 506.09 940.1 439.09 925.89 471.42 953.86 475.85 962.01 457.29 1333.44 516.12 1331.72 520.03 1330.75 519.87 1332.23 516.52 962.74 458 961.27 461.35 960.3 461.2 953.76 476.08 954.73 476.24 924.68 471.82 925 471.09 939.37 438.38 1366.54 505.69" fill="#361b70" strokeWidth="0px"/>
-                <path d="M961.26,461.35l121.97,19.32-6.39,14.54-.15.34-121.97-19.32-.97-.15,6.54-14.88.97.15ZM1075.97,494.84l6.05-13.77-121-19.16-6.05,13.77,121,19.16" fill="#361b70" strokeWidth="0px"/>
-                <polygon points="1076.85 495.21 1076.57 509.39 1076.42 509.73 1076.7 495.55 1076.85 495.21" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1076.7 495.55 1076.42 509.73 954.45 490.41 954.73 476.23 1076.7 495.55" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1212.34 501.71 1212.07 515.89 1206.16 529.34 1206.44 515.16 1212.34 501.71" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1324.6 533.88 1324.33 548.06 1206.16 529.34 1206.44 515.16 1324.6 533.88" fill="#535353" strokeWidth="0px"/>
-                <polygon points="1330.51 520.43 1324.6 533.88 1206.44 515.16 1212.34 501.71 1330.51 520.43" fill="#6f6f6f" strokeWidth="0px"/>
-                <path d="M1330.75,519.87l.97.15-6.08,13.84-.24.56-.07.16-120.1-19.02,6.39-14.56,119.13,18.87ZM1324.6,533.87l5.91-13.45-118.17-18.72-5.91,13.45,118.17,18.72" fill="#361b70" strokeWidth="0px"/>
-                <polygon points="1325.33 534.58 1325.05 548.76 1204.95 529.74 1205.22 515.56 1325.33 534.58" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1325.4 534.42 1325.13 548.6 1325.05 548.76 1325.33 534.58 1325.4 534.42" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1365.33 506.09 1351.34 537.93 1325.64 533.86 1331.72 520.03 1333.44 516.12 962.01 457.29 953.86 475.85 925.89 471.42 940.1 439.09 1365.33 506.09" fill="#9b9b9b" strokeWidth="0px"/>
-                <polygon points="1366.54 505.69 1366.27 519.87 1351.79 552.82 1352.07 538.64 1366.54 505.69" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1352.07 538.64 1351.79 552.82 1325.13 548.6 1325.4 534.42 1352.07 538.64" fill="#281454" strokeWidth="0px"/>
-                <polygon points="567.81 414.95 567.54 429.13 513.58 551.97 513.85 537.79 567.81 414.95" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1710.35 596.43 1710.08 610.61 954.38 490.58 954.66 476.4 1710.35 596.43" fill="#281454" strokeWidth="0px"/>
-                <polygon points="567.81 414.95 924.76 471.65 954.58 476.56 1710.35 596.43 1656.6 718.78 513.85 537.79 567.81 414.95" fill="#9b9b9b" strokeWidth="0px"/>
-                <polygon points="1656.6 718.78 1656.33 732.96 513.58 551.97 513.85 537.79 1656.6 718.78" fill="#747474" strokeWidth="0px"/>
-                <polygon points="1710.35 596.43 1710.08 610.61 1656.33 732.96 1656.6 718.78 1710.35 596.43" fill="#767676" strokeWidth="0px"/>
-                <polygon points="1657.33 719.49 1657.05 733.67 512.37 552.37 512.64 538.19 1657.33 719.49" fill="#281454" strokeWidth="0px"/>
-                <polygon points="567.09 414.24 925 471.09 567.81 414.95 513.85 537.79 1656.6 718.78 1710.35 596.43 954.66 476.4 1076.7 495.55 1711.56 596.03 1657.33 719.49 512.64 538.19 567.09 414.24" fill="#361b70" strokeWidth="0px"/>
-                <polygon points="1711.56 596.03 1711.29 610.21 1657.05 733.67 1657.33 719.49 1711.56 596.03" fill="#291455" strokeWidth="0px"/>
-            </g>
-            </g>
-            <g opacity=".17">
-            <g isolation="isolate">
-                <g clip-path="url(#clippath-9)">
-                <g clip-path="url(#clippath-10)">
-                    <path d="M536.16,517.76l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M584.04,525.34l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M631.91,532.93l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M679.79,540.51l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M727.66,548.09l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M775.54,555.67l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M823.42,563.26l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M871.29,570.84l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M919.17,578.42l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M967.04,586l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1014.91,593.59l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1062.78,601.17l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1110.66,608.75l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1158.53,616.34l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1206.41,623.92l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1254.28,631.5l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1302.16,639.08l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1350.03,646.67l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1397.91,654.25l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1445.78,661.83l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1493.66,669.41l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1541.53,677l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1589.41,684.58l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M544.12,499.64l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M592,507.22l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M639.87,514.8l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M687.75,522.39l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M735.62,529.97l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M783.5,537.55l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M831.38,545.14l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M879.25,552.72l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M927.13,560.3l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M975,567.88l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1022.88,575.47l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1070.75,583.05l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1118.62,590.63l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1166.49,598.21l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1214.37,605.8l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1262.24,613.38l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1310.12,620.96l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1357.99,628.54l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1405.87,636.13l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1453.74,643.71l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1501.62,651.29l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1549.49,658.87l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1597.37,666.46l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M552.08,481.51l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M599.96,489.09l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M647.83,496.68l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M695.71,504.26l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M743.58,511.84l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M791.46,519.42l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M839.34,527.01l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M887.21,534.59l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M935.09,542.17l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M982.96,549.75l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1030.84,557.34l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1078.71,564.92l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1126.59,572.5l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1174.46,580.09l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1222.34,587.67l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1270.21,595.25l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1318.09,602.83l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1365.96,610.42l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1413.83,618l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1461.7,625.58l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1509.58,633.16l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1557.45,640.75l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1605.33,648.33l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M560.05,463.39l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M607.92,470.97l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M655.79,478.55l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M703.67,486.14l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M751.54,493.72l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M799.42,501.3l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M847.3,508.89l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M895.17,516.47l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M943.05,524.05l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M990.92,531.63l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1038.8,539.22l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1086.67,546.8l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1134.55,554.38l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1182.42,561.96l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1230.3,569.55l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1278.17,577.13l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1326.05,584.71l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1373.92,592.29l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1421.8,599.88l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1469.67,607.46l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1517.54,615.04l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1565.41,622.62l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1613.29,630.21l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M568.01,445.27l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M615.89,452.85l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M663.76,460.43l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M711.63,468.01l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M759.5,475.6l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M807.38,483.18l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M855.26,490.76l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M903.13,498.34l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M951.01,505.93l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M998.88,513.51l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1046.76,521.09l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1094.63,528.67l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1142.51,536.26l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1190.38,543.84l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1238.26,551.42l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1286.13,559.01l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1334.01,566.59l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1381.88,574.17l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1429.76,581.75l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1477.63,589.34l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1525.51,596.92l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M1573.38,604.5l7.76-17.66,47.07,7.45-7.76,17.66-47.07-7.45M1621.25,612.08l7.76-17.66,47.06,7.45-7.76,17.66-47.06-7.45M582.09,411.24l-6.93,15.77-9.78-1.55-.2.47,9.78,1.55-7.76,17.66-9.78-1.55-.2.47,9.78,1.55-7.76,17.66-9.78-1.55-.2.47,9.78,1.55-7.76,17.66-9.78-1.55-.2.47,9.78,1.55-7.76,17.66-9.78-1.55-.2.46,9.78,1.55-7.76,17.66-9.78-1.55-.2.47,9.78,1.55-7.26,16.53.8.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.8.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.8.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,47.07,7.45-7.26,16.53.81.13,7.26-16.53,47.06,7.45-7.26,16.53.81.13,7.26-16.53,32.02,5.07.2-.47-32.02-5.07,7.76-17.66,32.02,5.07.2-.46-32.02-5.07,7.76-17.66,32.02,5.07.2-.47-32.02-5.07,7.76-17.66,32.02,5.07.2-.47-32.02-5.07,7.76-17.66,32.02,5.07.2-.47-32.02-5.07,7.76-17.66,32.02,5.07.2-.47-32.02-5.07,6.72-15.31-.81-.13-6.72,15.31-47.06-7.45,6.73-15.33-.81-.13-6.73,15.33-47.07-7.45,6.74-15.35-.81-.13-6.74,15.35-47.06-7.45,6.75-15.37-.81-.13-6.75,15.37-47.07-7.45,6.76-15.39-.81-.13-6.76,15.39-47.06-7.45,6.77-15.41-.81-.13-6.77,15.41-47.07-7.45,6.78-15.43-.81-.13-6.78,15.43-47.06-7.45,6.79-15.45-.81-.13-6.79,15.45-47.07-7.45,6.8-15.47-.81-.13-6.8,15.47-47.06-7.45,6.81-15.49-.81-.13-6.81,15.49-47.07-7.45,6.81-15.51-.81-.13-6.81,15.51-47.06-7.45,6.82-15.53-.81-.13-6.82,15.53-47.07-7.45,6.83-15.55-.81-.13-6.83,15.55-47.06-7.45,6.84-15.57-.81-.13-6.84,15.57-47.07-7.45,6.85-15.6-.81-.13-6.85,15.6-47.06-7.45,6.86-15.62-.81-.13-6.86,15.62-47.07-7.45,6.87-15.63-.81-.13-6.87,15.63-47.06-7.45,6.88-15.65-.81-.13-6.88,15.65-47.07-7.45,6.89-15.67-.81-.13-6.89,15.67-47.07-7.45,6.89-15.69-.8-.13-6.89,15.69-47.07-7.45,6.9-15.71-.81-.13-6.9,15.71-47.07-7.45,6.91-15.73-.8-.13-6.91,15.73-47.07-7.45,6.92-15.75-.81-.13-6.92,15.75-47.07-7.45,6.93-15.77-.8-.13M1067.95,472.86l-.81-.13-5.26,11.98-47.07-7.45,5.26-11.98-.81-.13-5.26,11.98-47.06-7.45,5.26-11.98-.81-.13-5.26,11.98-34.8-5.51-.2.47,34.8,5.51-.82,1.88.81.13.82-1.88,47.06,7.45-.82,1.88.81.13.82-1.88,47.07,7.45-.82,1.88.81.13.82-1.88,19.55,3.1.2-.47-19.55-3.1,5.26-11.98M1363.34,499.81l-.81-.13-5.45,12.41-47.07-7.45,5.44-12.38-.81-.13-5.44,12.38-47.06-7.45,5.42-12.34-.81-.13-5.42,12.34-47.07-7.45,5.4-12.3-.81-.13-5.4,12.3-47.06-7.45,5.39-12.27-.81-.13-5.39,12.27-47.07-7.45,5.37-12.23-.81-.13-5.37,12.23-47.06-7.45,5.36-12.19-.81-.13-5.36,12.19-47.07-7.45,5.34-12.16-.81-.13-5.34,12.16-47.06-7.45,5.32-12.12-.81-.13-5.32,12.12-34.8-5.51-.2.47,34.8,5.51-.78,1.77.81.13.78-1.77,47.06,7.45-.78,1.77.81.13.78-1.77,47.07,7.45-.78,1.77.81.13.78-1.77,47.06,7.45-.78,1.77.81.13.78-1.77,47.07,7.45-.78,1.77.81.13.78-1.77,47.06,7.45-.78,1.77.81.13.78-1.77,47.07,7.45-.78,1.77.81.13.78-1.77,47.06,7.45-.78,1.77.81.13.78-1.77,47.07,7.45-7.76,17.66-47.07-7.45,5.26-11.98-.81-.13-5.26,11.98-47.06-7.45,5.26-11.98-.81-.13-5.26,11.98-41.8-6.62-.2.47,41.8,6.62-.68,1.55.81.13.68-1.55,47.06,7.45-.68,1.55.81.13.68-1.55,47.07,7.45-.61,1.39.81.13.61-1.39,7.45,1.18.2-.47-7.45-1.18,7.76-17.66,7.45,1.18.2-.47-7.45-1.18,5.45-12.41" fill="#fff" strokeWidth="0px"/>
-                </g>
-                </g>
-            </g>
-            </g>
-            <g id="VIEWING_ROOM" data-name="VIEWING ROOM">
-            <g isolation="isolate">
-                <polygon points="668.12 423.06 667.84 437.44 565.59 421.25 565.87 406.86 668.12 423.06" fill="#281454" strokeWidth="0px"/>
-                <polygon points="668.12 423.06 667.84 437.44 656.11 464.14 656.39 449.75 668.12 423.06" fill="#00a7a7" strokeWidth="0px"/>
-                <polygon points="669.33 422.65 669.05 437.04 657.32 463.73 657.6 449.35 669.33 422.65" fill="#291455" strokeWidth="0px"/>
-                <polygon points="679.26 453.37 678.98 467.76 656.11 464.14 656.39 449.75 679.26 453.37" fill="#281454" strokeWidth="0px"/>
-                <polygon points="565.87 406.86 565.59 421.25 513.92 538.87 514.19 524.49 565.87 406.86" fill="#291455" strokeWidth="0px"/>
-                <polygon points="639.31 544.31 639.04 558.69 513.92 538.87 514.19 524.49 639.31 544.31" fill="#00a6a6" strokeWidth="0px"/>
-                <polygon points="679.26 453.37 678.98 467.76 639.04 558.69 639.31 544.31 679.26 453.37" fill="#00a7a7" strokeWidth="0px"/>
-                <path d="M657.6,449.35l22.87,3.62-40.43,92.04-127.06-20.12,52.16-118.74,104.19,16.5-11.73,26.7ZM639.31,544.31l39.95-90.93-22.87-3.62,11.73-26.7-102.25-16.19-51.67,117.63,125.12,19.82" fill="#361b70" strokeWidth="0px"/>
-                <polygon points="668.12 423.06 656.39 449.75 679.26 453.37 639.31 544.31 514.19 524.49 565.87 406.86 668.12 423.06" fill="#00dcdc" strokeWidth="0px"/>
-                <polygon points="680.47 452.97 680.19 467.36 639.76 559.4 640.04 545.02 680.47 452.97" fill="#291455" strokeWidth="0px"/>
-                <polygon points="640.04 545.02 639.76 559.4 512.7 539.28 512.98 524.89 640.04 545.02" fill="#281454" strokeWidth="0px"/>
-            </g>
-            </g>
-            <g isolation="isolate">
-            <polygon points="753.71 465.28 753.44 479.45 680.3 467.86 680.57 453.69 753.71 465.28" fill="#281454" strokeWidth="0px"/>
-            <polygon points="680.57 453.69 680.3 467.86 657.08 522.1 657.35 507.93 680.57 453.69" fill="#291454" strokeWidth="0px"/>
-            <polygon points="729.92 519.43 729.65 533.59 657.08 522.1 657.35 507.93 729.92 519.43" fill="#16764b" strokeWidth="0px"/>
-            <polygon points="753.71 465.28 753.44 479.45 729.65 533.59 729.92 519.43 753.71 465.28" fill="#16774c" strokeWidth="0px"/>
-            <polygon points="753.71 465.28 729.92 519.43 657.35 507.93 680.57 453.69 753.71 465.28" fill="#1d9d63" strokeWidth="0px"/>
-            <polygon points="730.65 520.13 730.37 534.3 655.86 522.5 656.14 508.33 730.65 520.13" fill="#281454" strokeWidth="0px"/>
-            <polygon points="754.92 464.88 754.65 479.04 730.37 534.3 730.65 520.13 754.92 464.88" fill="#291455" strokeWidth="0px"/>
-            <path d="M679.85,452.98l75.08,11.89-24.27,55.26-74.51-11.8,23.71-55.35ZM729.92,519.43l23.79-54.15-73.14-11.58-23.22,54.24,72.57,11.49" fill="#361b70" strokeWidth="0px"/>
-            </g>
-            <g isolation="isolate">
-            <polygon points="729.81 520.13 729.54 534.31 656.84 522.8 657.12 508.62 729.81 520.13" fill="#281454" strokeWidth="0px"/>
-            <polygon points="657.12 508.62 656.84 522.8 641.07 558.71 641.34 544.53 657.12 508.62" fill="#291455" strokeWidth="0px"/>
-            <polygon points="714.04 556.04 713.76 570.22 641.07 558.71 641.34 544.53 714.04 556.04" fill="#a64594" strokeWidth="0px"/>
-            <polygon points="729.81 520.13 729.54 534.31 713.76 570.22 714.04 556.04 729.81 520.13" fill="#a74695" strokeWidth="0px"/>
-            <polygon points="729.81 520.13 714.04 556.04 641.34 544.53 657.12 508.62 729.81 520.13" fill="#dc5cc4" strokeWidth="0px"/>
-            <polygon points="714.76 556.75 714.49 570.93 639.86 559.11 640.13 544.93 714.76 556.75" fill="#281454" strokeWidth="0px"/>
-            <polygon points="731.02 519.73 730.75 533.91 714.49 570.93 714.76 556.75 731.02 519.73" fill="#291455" strokeWidth="0px"/>
-            <path d="M656.39,507.91l74.63,11.82-16.26,37.02-74.63-11.82,16.26-37.02ZM714.04,556.04l15.77-35.91-72.7-11.51-15.77,35.91,72.7,11.51" fill="#361b70" strokeWidth="0px"/>
-            </g>
-            <g isolation="isolate">
-            <polygon points="1065.47 514.77 1065.19 528.95 755.18 479.85 755.45 465.67 1065.47 514.77" fill="#281454" strokeWidth="0px"/>
-            <polygon points="755.45 465.67 755.18 479.85 715.17 570.92 715.45 556.74 755.45 465.67" fill="#291455" strokeWidth="0px"/>
-            <polygon points="1025.46 605.84 1025.19 620.02 715.17 570.92 715.45 556.74 1025.46 605.84" fill="#a67b48" strokeWidth="0px"/>
-            <polygon points="1065.47 514.77 1065.19 528.95 1025.19 620.02 1025.46 605.84 1065.47 514.77" fill="#a77c49" strokeWidth="0px"/>
-            <path d="M754.73,464.96l311.95,49.41-40.5,92.19-311.95-49.41,40.5-92.19ZM1025.46,605.84l40.01-91.07-310.01-49.1-40.01,91.07,310.01,49.1" fill="#361b70" strokeWidth="0px"/>
-            <polygon points="1065.47 514.77 1025.46 605.84 715.45 556.74 755.45 465.67 1065.47 514.77" fill="#dca360" strokeWidth="0px"/>
-            <polygon points="1026.19 606.55 1025.91 620.73 713.96 571.33 714.24 557.14 1026.19 606.55" fill="#281454" strokeWidth="0px"/>
-            <polygon points="1066.68 514.37 1066.41 528.55 1025.91 620.73 1026.19 606.55 1066.68 514.37" fill="#291455" strokeWidth="0px"/>
-            </g>
-            <g id="READING_AREa" data-name="READING AREa">
-            <g isolation="isolate">
-                <polygon points="1163.63 529.98 1163.35 544.16 1153.64 566.19 1153.91 552.01 1163.63 529.98" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1153.91 552.01 1153.64 566.19 1057.4 550.94 1057.68 536.76 1153.91 552.01" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1418.65 570.37 1418.37 584.55 1163.35 544.16 1163.63 529.98 1418.65 570.37" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1057.68 536.76 1057.4 550.94 1026.95 620.16 1027.22 605.98 1057.68 536.76" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1418.65 570.37 1378.56 661.63 1027.22 605.98 1057.68 536.76 1153.91 552.01 1163.63 529.98 1418.65 570.37" fill="#8edc89" strokeWidth="0px"/>
-                <polygon points="1378.56 661.63 1378.29 675.81 1026.95 620.16 1027.22 605.98 1378.56 661.63" fill="#6ba667" strokeWidth="0px"/>
-                <polygon points="1418.65 570.37 1418.38 584.55 1378.29 675.81 1378.56 661.63 1418.65 570.37" fill="#6ca768" strokeWidth="0px"/>
-                <path d="M1162.9,529.27l256.96,40.7-40.58,92.37-353.28-55.95,30.94-70.33,96.24,15.24,9.71-22.03ZM1378.56,661.63l40.09-91.26-255.02-40.39-9.71,22.03-96.24-15.24-30.45,69.22,351.34,55.65" fill="#361b70" strokeWidth="0px"/>
-                <polygon points="1379.29 662.34 1379.01 676.52 1025.74 620.57 1026.01 606.39 1379.29 662.34" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1419.86 569.97 1419.59 584.15 1379.01 676.52 1379.29 662.34 1419.86 569.97" fill="#291455" strokeWidth="0px"/>
-            </g>
-            </g>
-            <g id="E-Lib">
-            <g isolation="isolate">
-                <polygon points="1577.39 595.7 1577.12 609.76 1419.5 584.8 1419.77 570.73 1577.39 595.7" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1419.77 570.73 1419.5 584.8 1403.78 620.59 1404.05 606.52 1419.77 570.73" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1561.67 631.49 1561.4 645.55 1403.78 620.59 1404.05 606.52 1561.67 631.49" fill="#007ea6" strokeWidth="0px"/>
-                <polygon points="1577.39 595.7 1577.12 609.76 1561.4 645.55 1561.67 631.49 1577.39 595.7" fill="#007fa7" strokeWidth="0px"/>
-                <polygon points="1577.39 595.7 1561.67 631.49 1404.05 606.52 1419.77 570.73 1577.39 595.7" fill="#00a8dc" strokeWidth="0px"/>
-                <polygon points="1562.4 632.2 1562.13 646.27 1402.55 620.99 1402.83 606.93 1562.4 632.2" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1578.62 595.3 1578.34 609.36 1562.13 646.27 1562.4 632.2 1578.62 595.3" fill="#291455" strokeWidth="0px"/>
-                <path d="M1419.04,570.02l159.58,25.27-16.21,36.91-159.58-25.27,16.21-36.91ZM1561.67,631.49l15.72-35.79-157.62-24.97-15.72,35.79,157.62,24.97" fill="#361b70" strokeWidth="0px"/>
-            </g>
-            </g>
-            <g id="HALLWAY">
-            <g isolation="isolate">
-                <polygon points="1404.08 606.75 1403.81 620.93 1395.63 639.55 1395.9 625.37 1404.08 606.75" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1567.32 632.61 1567.05 646.79 1403.81 620.93 1404.08 606.75 1567.32 632.61" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1559.15 651.23 1558.87 665.41 1395.63 639.55 1395.9 625.37 1559.15 651.23" fill="#9d9d9d" strokeWidth="0px"/>
-                <polygon points="1567.32 632.61 1567.05 646.79 1558.87 665.41 1559.15 651.23 1567.32 632.61" fill="#9f9f9f" strokeWidth="0px"/>
-                <polygon points="1567.32 632.61 1559.15 651.23 1395.9 625.37 1404.08 606.75 1567.32 632.61" fill="#d1d1d1" strokeWidth="0px"/>
-                <polygon points="1559.87 651.94 1559.6 666.12 1394.41 639.95 1394.69 625.77 1559.87 651.94" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1568.54 632.2 1568.26 646.38 1559.6 666.12 1559.87 651.94 1568.54 632.2" fill="#291455" strokeWidth="0px"/>
-                <path d="M1403.36,606.04l165.18,26.16-8.67,19.73-165.18-26.16,8.67-19.73ZM1559.15,651.23l8.18-18.62-163.24-25.86-8.18,18.62,163.24,25.86" fill="#361b70" strokeWidth="0px"/>
-            </g>
-            </g>
-            <g id="CONFERENCE">
-            <g isolation="isolate">
-                <polygon points="1553.3 650.41 1553.03 664.59 1395.39 639.63 1395.67 625.45 1553.3 650.41" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1395.67 625.45 1395.39 639.63 1379.48 675.86 1379.75 661.68 1395.67 625.45" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1537.38 686.65 1537.11 700.83 1379.48 675.86 1379.75 661.68 1537.38 686.65" fill="#7467a6" strokeWidth="0px"/>
-                <polygon points="1553.3 650.41 1553.03 664.59 1537.11 700.83 1537.38 686.65 1553.3 650.41" fill="#7568a7" strokeWidth="0px"/>
-                <polygon points="1553.3 650.41 1537.38 686.65 1379.75 661.68 1395.67 625.45 1553.3 650.41" fill="#9a89dc" strokeWidth="0px"/>
-                <polygon points="1538.11 687.36 1537.83 701.54 1378.26 676.26 1378.54 662.08 1538.11 687.36" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1554.51 650.01 1554.24 664.19 1537.83 701.54 1538.11 687.36 1554.51 650.01" fill="#291455" strokeWidth="0px"/>
-                <path d="M1394.94,624.74l159.57,25.27-16.41,37.35-159.57-25.27,16.41-37.35ZM1537.38,686.65l15.92-36.24-157.63-24.97-15.92,36.24,157.63,24.97" fill="#361b70" strokeWidth="0px"/>
-            </g>
-            </g>
-            <g id="THETSIS_AND_DISSERTATION_ROOM" data-name="THETSIS AND DISSERTATION ROOM">
-            <g isolation="isolate">
-                <polygon points="1708.78 587.73 1708.51 601.91 1611.05 586.48 1611.32 572.29 1708.78 587.73" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1611.32 572.29 1611.05 586.48 1599.24 613.35 1599.52 599.17 1611.32 572.29" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1599.52 599.17 1599.24 613.35 1575.99 609.67 1576.26 595.49 1599.52 599.17" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1576.26 595.49 1575.99 609.67 1536.41 699.77 1536.68 685.59 1576.26 595.49" fill="#291455" strokeWidth="0px"/>
-                <polygon points="1708.78 587.73 1657.4 704.71 1536.68 685.59 1576.26 595.49 1599.52 599.17 1611.32 572.29 1708.78 587.73" fill="#dc5e69" strokeWidth="0px"/>
-                <polygon points="1657.4 704.71 1657.12 718.89 1536.41 699.77 1536.68 685.59 1657.4 704.71" fill="#a6474f" strokeWidth="0px"/>
-                <polygon points="1708.78 587.73 1708.51 601.91 1657.12 718.89 1657.4 704.71 1708.78 587.73" fill="#a74850" strokeWidth="0px"/>
-                <path d="M1610.6,571.59l99.41,15.74-51.87,118.09-122.67-19.43,40.07-91.21,23.26,3.68,11.81-26.88ZM1657.4,704.71l51.39-116.98-97.46-15.44-11.81,26.88-23.25-3.68-39.58,90.1,120.71,19.12" fill="#361b70" strokeWidth="0px"/>
-                <polygon points="1658.13 705.42 1657.86 719.6 1535.19 700.17 1535.46 685.99 1658.13 705.42" fill="#281454" strokeWidth="0px"/>
-                <polygon points="1710.01 587.33 1709.73 601.51 1657.86 719.6 1658.13 705.42 1710.01 587.33" fill="#291455" strokeWidth="0px"/>
-            </g>
-            </g>
-        </g>
-        </svg>
-    </div>
-  )
-}
-
-export default NewAdminFloor3;
