+++ src/api/new admin/floors/floors/NewAdminFloor2.jsx
@@ -1,0 +1,582 @@
+import React from 'react'
+import { useContext, useState } from "react";
+
+export const NewAdminFloor2 = () => {
+
+    const [showInfoPanel, setShowInfoPanel] = useState(false);
+    const [roomInfo, setRoomInfo] = useState(null);
+    const [emoji, setEmoji] = useState("👷‍♂️");
+
+    const Floor2 = {
+    "2": {
+    "rooms": {
+        
+        }
+    }
+
+  }
+
+    const roomClicked = (key) =>{
+    setRoomInfo({ key, ...Floor1["1"].rooms[key] });
+    console.log(roomInfo);
+  } 
+
+    const roomOfs = -5;
+
+    const roomOnMouseEnter = (e, room, roomText) =>{
+    e.currentTarget.setAttribute('fill','#FFFFFF');
+    e.currentTarget.style.transform = `translate(${roomOfs}px, ${roomOfs}px)`; 
+    document.querySelector(roomText).style.transform = `translate(${roomOfs}px, ${roomOfs}px)`;
+    const roomStroke = document.querySelectorAll(room); 
+    console.log(roomStroke)
+    roomStroke.forEach(ro => {
+            ro.style.transform = `translate(${roomOfs}px, ${roomOfs}px)`;
+    });
+  }
+    const roomOnMouseLeave = (e, room, roomText, roomColor) =>{
+    e.currentTarget.setAttribute('fill', roomColor);
+    e.currentTarget.style.transform = 'translate(0px, 0px)'; 
+    document.querySelector(roomText).style.transform = 'translate(0px, 0px)'; 
+    const roomStroke = document.querySelectorAll(room); 
+    roomStroke.forEach(ro => {
+            ro.style.transform = 'translate(0px, 0px)'; 
+    });
+  }
+
+  return (
+    <div className='-mt-32 lg:flex w-screen justify-center items-center'>
+        <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 2420.06 1051.82">
+        <defs>
+            <clipPath id="clippath">
+            <polygon points="512.18 695.52 614.08 470.33 1791.29 651.34 1689.4 876.52 512.18 695.52" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-2">
+            <polygon points="512.18 695.52 614.08 470.33 1791.29 651.34 1689.4 876.52 512.18 695.52" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-3">
+            <polyline points="594.1 514.49 530.16 655.77 936.4 718.24 920.7 752.92 1030.98 769.88 1028.69 774.93 1171.95 796.96 1189.94 757.23 1707.38 836.79 1771.02 696.13 1625.87 673.81 1630.44 663.71 1525.82 647.63 1521.26 657.72 1408.15 640.33 1428.42 595.55 976.95 526.12 956.4 571.54 843.3 554.15 849.01 541.53 844.86 540.9 752.41 526.68 748.16 526.03 743.02 537.38 594.1 514.49" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-4">
+            <polygon points="512.18 695.52 614.08 470.33 1791.29 651.34 1689.4 876.52 512.18 695.52" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-6">
+            <path d="M585.64,476.36l-.27,14.18c-19.66,43.26-56.26,124.61-56.95,129.24l.27-14.18c.69-4.63,37.29-85.98,56.95-129.24Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-7">
+            <path d="M831.2,667.17c-264.99-41.97-298.31-46.92-302.78-47.39l.27-14.18c4.48.48,37.79,5.42,302.78,47.39-.09,4.72-.18,9.43-.27,14.15.09-4.72.18-9.43.27-14.15,1.04.16,2.08.33,3.13.5-.04,2.28-.09,4.57-.13,6.85-.04,2.29-.09,4.58-.13,6.87,0,.15,0,.3,0,.46-1.05-.17-2.09-.33-3.13-.5Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-8">
+            <path d="M834.33,667.67c.09-4.73.18-9.45.27-14.18,75,11.88,168.36,26.69,284.61,45.17-.09,4.72-.18,9.43-.27,14.15.09-4.72.18-9.43.27-14.15.66.1,1.31.21,1.97.31-.09,4.72-.18,9.43-.27,14.15.09-4.72.18-9.43.27-14.15,284.99,45.29,572.68,91.15,586.61,93.37l-.27,14.15v.03c-13.96-2.23-302.94-48.29-588.57-93.68-116.24-18.47-209.61-33.29-284.61-45.17Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-9">
+            <path d="M528.32,605.54l-.27,14.18s-.1-.01-.14-.01l.27-14.18s.1,0,.14.01Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-10">
+            <path d="M528.18,605.53l-.27,14.18c-.07,0-.14,0-.21.03l.27-14.18c.06-.02.14-.03.21-.03Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-11">
+            <path d="M527.98,619.74c.09-4.73.18-9.45.27-14.18.05,0,.1,0,.16.01l-.27,14.18c-.06,0-.11-.01-.16-.01Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-12">
+            <path d="M527.74,619.75l.27-14.18c0,.45.17.51.36.56l-.27,14.18c-.19-.05-.37-.11-.36-.56Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-13">
+            <path d="M527.97,619.73c.09-4.73.18-9.45.27-14.18,0,0,0,0,.01,0-.09,4.73-.18,9.45-.27,14.18,0,0,0,0-.01,0Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-14">
+            <path d="M527.97,619.73s-.05,0-.07,0l.27-14.18s.04,0,.07,0c-.09,4.73-.18,9.45-.27,14.18Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-15">
+            <path d="M1708.54,793.08l-.27,14.18c-8.33-1.33-528.57-84.27-875.54-139.22-170.69-27.04-299.38-47.28-304.62-47.73l.27-14.18c5.24.44,133.92,20.69,304.62,47.73,346.97,54.95,867.21,137.89,875.54,139.22Z" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-16">
+            <polygon points="528.27 603.76 601.43 437.22 1782.18 624.24 1709.02 790.78 528.27 603.76" fill="none" strokeWidth="0px"/>
+            </clipPath>
+            <clipPath id="clippath-17">
+            <path d="M1394.06,602.53l16.33-37.18-438.62-69.47-16.22,36.92-369.41-58.74s-59.41,130.7-57.04,129.84c2.37-.87,1179.92,186.88,1179.92,186.88l56.73-129.13-371.69-59.11" fill="none" strokeWidth="0px"/>
+            </clipPath>
+        </defs>
+        <g id="Layer_4" data-name="Layer 4">
+            <path d="M899.85,430.34h408.68c103.66,0,187.83,84.16,187.83,187.83v140.28c0,18.29-14.85,33.13-33.13,33.13h-340.08c-141.53,0-256.43-114.9-256.43-256.43v-71.68c0-18.29,14.85-33.13,33.13-33.13Z" fill="#00e5ff" opacity=".4" strokeWidth="0px"/>
+            <path d="M1466.53,424.21l104.44,78.51,105.92-74.75c9.46-6.68,22.01,2.76,18.32,13.77l-41.37,123.27,104.44,78.51c9.33,7.01,4.36,21.83-7.26,21.62l-130.01-2.32-41.37,123.27c-3.69,11.01-19.32,10.73-22.8-.41l-38.98-124.71-130.01-2.32c-11.61-.21-16.3-15.19-6.84-21.87l105.92-74.75-38.98-124.71c-3.48-11.14,9.25-20.12,18.58-13.11Z" fill="#ff74d2" opacity=".32" strokeWidth="0px"/>
+            <circle cx="793.66" cy="561.32" r="223.04" fill="#ffff85" opacity=".4" strokeWidth="0px"/>
+        </g>
+        <g id="old_floor" data-name="old floor">
+            <g>
+            <g id="floor">
+                <g>
+                <polygon points="747.96 526.47 747.72 538.16 742.59 549.51 742.82 537.81 747.96 526.47" fill="#212121" strokeWidth="0px"/>
+                <polygon points="742.82 537.81 742.59 549.51 593.67 526.61 593.9 514.92 742.82 537.81" fill="#212121" strokeWidth="0px"/>
+                <polygon points="848.13 541.87 847.9 553.56 747.72 538.16 747.96 526.47 848.13 541.87" fill="#212121" strokeWidth="0px"/>
+                <polygon points="848.13 541.87 847.9 553.56 842.19 566.18 842.43 554.49 848.13 541.87" fill="#767676" strokeWidth="0px"/>
+                <polygon points="848.97 541.6 848.74 553.29 843.03 565.91 843.27 554.21 848.97 541.6" fill="#212121" strokeWidth="0px"/>
+                <polygon points="976.75 526.56 976.51 538.25 955.96 583.67 956.19 571.97 976.75 526.56" fill="#212121" strokeWidth="0px"/>
+                <polygon points="956.19 571.97 955.96 583.67 842.19 566.17 842.42 554.48 956.19 571.97" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1427.54 595.88 1427.3 607.57 976.51 538.25 976.75 526.56 1427.54 595.88" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1427.54 595.88 1427.3 607.57 1407.05 652.35 1407.28 640.66 1427.54 595.88" fill="#767676" strokeWidth="0px"/>
+                <polygon points="1428.37 595.61 1428.14 607.3 1407.88 652.09 1408.11 640.4 1428.37 595.61" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1521.05 658.16 1520.81 669.85 1407.05 652.35 1407.28 640.66 1521.05 658.16" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1525.61 648.06 1525.38 659.76 1520.81 669.85 1521.05 658.16 1525.61 648.06" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1629.56 664.05 1629.33 675.74 1525.38 659.76 1525.61 648.06 1629.56 664.05" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1629.57 664.05 1629.33 675.74 1624.77 685.84 1625 674.14 1629.57 664.05" fill="#767676" strokeWidth="0px"/>
+                <polygon points="1630.4 663.78 1630.16 675.47 1625.6 685.57 1625.83 673.87 1630.4 663.78" fill="#212121" strokeWidth="0px"/>
+                <polygon points="593.9 514.92 593.67 526.61 529.9 667.53 530.14 655.84 593.9 514.92" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1770.14 696.46 1769.91 708.15 1624.76 685.83 1625 674.14 1770.14 696.46" fill="#212121" strokeWidth="0px"/>
+                <polygon points="936.36 718.3 936.13 729.99 529.9 667.53 530.14 655.84 936.36 718.3" fill="#747474" strokeWidth="0px"/>
+                <polygon points="935.52 718.57 935.28 730.26 529.06 667.8 529.29 656.1 935.52 718.57" fill="#212121" strokeWidth="0px"/>
+                <polygon points="936.36 718.3 936.12 729.99 920.43 764.68 920.67 752.99 936.36 718.3" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1030.94 769.95 1030.7 781.64 920.43 764.68 920.67 752.99 1030.94 769.95" fill="#747474" strokeWidth="0px"/>
+                <polygon points="1030.1 770.21 1029.86 781.91 919.59 764.95 919.82 753.26 1030.1 770.21" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1030.94 769.95 1030.7 781.64 1028.42 786.69 1028.66 775 1030.94 769.95" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1189.23 757.19 1188.99 768.88 1171.01 808.61 1171.25 796.92 1189.23 757.19" fill="#767676" strokeWidth="0px"/>
+                <polygon points="1171.25 796.92 1171.01 808.61 1028.43 786.69 1028.66 775 1171.25 796.92" fill="#747474" strokeWidth="0px"/>
+                <polygon points="1706.67 836.75 1706.43 848.44 1188.99 768.88 1189.23 757.19 1706.67 836.75" fill="#747474" strokeWidth="0px"/>
+                <polygon points="1189.72 757.66 1189.49 769.35 1171.51 809.08 1171.74 797.39 1189.72 757.66" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1171.74 797.39 1171.51 809.08 1027.58 786.95 1027.82 775.26 1171.74 797.39" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1770.15 696.46 1769.91 708.15 1706.43 848.44 1706.67 836.75 1770.15 696.46" fill="#767676" strokeWidth="0px"/>
+                <polygon points="1427.54 595.88 1407.28 640.66 1521.05 658.16 1525.62 648.06 1629.57 664.05 1625 674.14 1770.15 696.46 1706.67 836.75 1189.23 757.19 1171.25 796.92 1028.66 775 1030.94 769.95 920.67 752.99 936.36 718.3 530.14 655.84 593.9 514.92 742.82 537.81 747.96 526.47 848.13 541.87 842.43 554.49 956.2 571.97 976.75 526.57 1427.54 595.88" fill="#9b9b9b" strokeWidth="0px"/>
+                <polygon points="1707.17 837.23 1706.94 848.92 1189.49 769.35 1189.72 757.66 1707.17 837.23" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1770.98 696.19 1770.75 707.89 1706.94 848.92 1707.17 837.23 1770.98 696.19" fill="#212121" strokeWidth="0px"/>
+                <path d="M1625.83,673.87l145.15,22.32-63.82,141.03-517.45-79.57-17.98,39.74-143.93-22.13,2.28-5.05-110.28-16.95,15.7-34.69-406.22-62.46,64.1-141.67,148.92,22.9,5.14-11.35,101.53,15.61-5.71,12.62,112.42,17.29,20.55-45.42,452.14,69.53-20.26,44.78,112.43,17.29,4.57-10.09,105.29,16.19-4.57,10.09h0ZM1706.66,836.75l63.48-140.29-145.15-22.32,4.57-10.09-103.95-15.98-4.57,10.09-113.78-17.5,20.26-44.78-450.79-69.32-20.55,45.42-113.78-17.5,5.71-12.62-100.18-15.4-5.14,11.35-148.92-22.9-63.77,140.92,406.22,62.46-15.7,34.69,110.28,16.95-2.28,5.05,142.58,21.92,17.98-39.74,517.44,79.56" fill="#2c2c2c" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="grid-2" data-name="grid" opacity=".17">
+                <g clip-path="url(#clippath-2)">
+                <g clip-path="url(#clippath-3)">
+                    <path d="M928.11,749.31l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M976.69,756.78l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1025.28,764.25l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1073.87,771.72l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1122.44,779.19l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M936.18,731.45l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M984.77,738.93l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1033.35,746.4l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1081.94,753.87l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1130.52,761.34l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M555.59,653.84l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M604.17,661.31l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M652.76,668.78l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M701.34,676.25l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M749.93,683.72l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M798.51,691.19l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M847.1,698.66l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M895.68,706.13l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M944.26,713.6l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M992.85,721.07l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1041.43,728.54l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1090.02,736.01l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1138.6,743.48l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1187.19,750.95l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1235.77,758.42l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1284.35,765.89l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1332.94,773.36l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1381.53,780.84l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1430.11,788.31l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1478.7,795.78l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1527.29,803.25l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1575.87,810.72l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1624.45,818.19l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M563.66,635.98l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M612.24,643.45l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M660.83,650.93l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M709.42,658.4l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M758,665.87l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M806.59,673.34l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M855.17,680.81l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M903.76,688.28l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M952.34,695.75l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1000.93,703.22l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1049.52,710.69l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1098.1,718.16l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1146.68,725.63l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1195.27,733.1l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1243.86,740.57l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1292.43,748.04l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1341.02,755.51l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1389.61,762.98l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1438.19,770.45l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1486.77,777.92l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1535.36,785.39l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1583.95,792.86l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1632.53,800.33l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M571.75,618.12l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M620.33,625.6l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M668.91,633.07l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M717.5,640.54l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M766.09,648.01l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M814.66,655.48l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M863.25,662.95l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M911.84,670.42l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M960.42,677.89l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1009,685.36l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1057.59,692.83l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1106.18,700.3l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1154.76,707.77l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1203.35,715.24l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1251.94,722.71l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1300.52,730.18l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1349.1,737.65l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1397.69,745.12l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1446.28,752.6l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1494.86,760.06l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1543.44,767.54l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1592.03,775.01l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1640.61,782.48l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M579.82,600.27l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M628.41,607.74l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M676.99,615.21l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M725.58,622.68l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M774.17,630.15l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M822.75,637.63l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M871.33,645.1l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M919.92,652.57l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M968.5,660.04l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1017.09,667.51l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1065.67,674.98l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1114.26,682.45l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1162.84,689.92l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1211.42,697.39l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1260.01,704.86l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1308.59,712.33l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1357.18,719.8l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1405.76,727.27l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1454.35,734.74l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1502.94,742.21l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1551.52,749.68l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1600.11,757.15l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1648.69,764.62l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M587.9,582.41l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M636.48,589.88l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M685.07,597.35l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M733.65,604.82l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M782.24,612.3l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M830.82,619.77l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M879.41,627.24l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M928,634.71l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M976.58,642.19l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1025.17,649.66l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1073.75,657.13l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1122.34,664.6l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1170.92,672.07l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1219.51,679.54l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1268.09,687.01l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1316.67,694.48l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1365.26,701.95l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1413.84,709.42l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1462.43,716.89l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1511.01,724.36l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1559.6,731.83l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1608.18,739.3l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1656.76,746.77l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M595.99,564.56l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M644.56,572.03l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M693.15,579.5l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M741.74,586.97l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M790.32,594.44l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M838.9,601.91l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M887.49,609.38l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M936.08,616.85l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M984.65,624.33l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1033.24,631.8l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1081.83,639.27l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1130.41,646.74l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1179,654.21l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1227.59,661.68l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1276.18,669.15l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1324.75,676.62l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1373.34,684.09l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1421.93,691.56l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1470.52,699.03l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1519.09,706.5l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1567.68,713.97l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1616.27,721.44l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1664.85,728.91l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M604.06,546.7l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M652.65,554.17l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M701.23,561.64l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M749.82,569.11l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M798.41,576.58l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M846.98,584.05l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M895.57,591.52l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M944.16,599l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M992.74,606.47l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1041.32,613.94l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1089.91,621.41l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1138.5,628.89l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1187.07,636.36l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1235.66,643.83l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1284.25,651.3l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1332.83,658.77l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1381.41,666.24l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1430,673.71l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1478.59,681.18l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1527.17,688.65l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1575.76,696.12l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1624.35,703.59l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1672.93,711.06l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M757.89,551.26l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M952.24,581.14l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1000.82,588.61l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1049.41,596.08l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1097.99,603.55l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1146.58,611.03l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1195.16,618.5l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1243.75,625.97l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1292.33,633.44l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1340.91,640.91l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1535.25,670.79l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M960.31,563.28l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1008.89,570.75l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1057.48,578.22l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1106.07,585.7l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1154.65,593.17l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1203.24,600.64l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1251.83,608.11l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1300.41,615.59l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1348.99,623.06l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M968.4,545.43l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1016.97,552.9l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1065.56,560.37l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1114.15,567.84l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1162.73,575.31l7.87-17.4,47.76,7.35-7.87,17.4-47.76-7.35M1211.31,582.78l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1259.9,590.25l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1308.49,597.73l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1357.06,605.2l7.87-17.4,47.77,7.35-7.87,17.4-47.77-7.35M1414.38,593.39l-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.76-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.82-.12-.65,1.43-47.77-7.35.65-1.43-.17-.02-20.55,45.42-.65-.1-4.33,9.57-47.77-7.35,4.33-9.57-.82-.12-4.33,9.57-47.77-7.35,4.33-9.57-.82-.12-4.33,9.57-47.76-7.35,7.87-17.4,32.49,5,.2-.46-32.49-5,1.96-4.34-.82-.12-1.96,4.34-47.77-7.35,1.96-4.34-.82-.12-1.96,4.34-18.96-2.92-.2.46,18.96,2.92-7.87,17.4-47.77-7.35,4.9-10.83-.82-.12-4.9,10.83-47.77-7.35,4.9-10.83-.82-.12-4.9,10.83-47.77-7.35,4.9-10.83-.82-.12-4.9,10.83-22.12-3.41-.2.46,22.12,3.41-7.87,17.4-22.12-3.41-.2.46,22.12,3.41-7.87,17.4-22.12-3.41-.2.46,22.12,3.41-7.87,17.4-22.12-3.41-.2.45,22.12,3.41-7.87,17.4-22.12-3.41-.2.46,22.12,3.41-7.87,17.4-22.12-3.41-.2.46,22.12,3.41-7.87,17.4-22.12-3.41-.2.45,22.12,3.41-7.87,17.4-22.12-3.41-.2.46,22.12,3.41-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.76,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-7.87,17.4-4.58-.7-.2.46,4.58.7-7.87,17.4-4.58-.7-.2.46,4.58.7-1.8,3.98.82.12,1.8-3.98,47.77,7.35-1.8,3.98.82.12,1.8-3.98,47.77,7.35-1.8,3.98.82.12,1.8-3.98,47.77,7.35-4.09,9.04.82.12,4.09-9.04,47.76,7.35-4.09,9.04.82.12,4.09-9.04,47.77,7.35-4.09,9.04.82.12,4.09-9.04,5.21.8.2-.46-5.21-.8,7.87-17.4,5.21.8.2-.46-5.21-.8,7.87-17.4,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.76,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,47.76,7.35-2.27,5.01.82.12,2.27-5.01,47.77,7.35-2.27,5.01.82.12,2.27-5.01,36.81,5.66.2-.46-36.81-5.66,7.87-17.4,36.81,5.66.2-.45-36.81-5.66,7.87-17.4,36.81,5.66.2-.46-36.81-5.66,7.87-17.4,36.81,5.66.2-.46-36.81-5.66,7.87-17.4,36.81,5.66.2-.45-36.81-5.66,7.87-17.4,36.81,5.66.2-.46-36.81-5.66,7.87-17.4,36.81,5.66.2-.46-36.81-5.66,7.87-17.4,36.81,5.66.2-.46-36.81-5.66,4.62-10.2-.82-.12-4.62,10.2-47.77-7.35,4.62-10.2-.82-.12-4.62,10.2-47.76-7.35,4.62-10.2-.82-.12-4.62,10.2-47.77-7.35,7.87-17.4,37.42,5.76.2-.46-37.42-5.76,1.1-2.44-.82-.12-1.1,2.44-47.77-7.35,1.1-2.44-.82-.12-1.1,2.44-17.79-2.74-.2.46,17.79,2.74-7.87,17.4-47.76-7.35,4.62-10.2-.82-.12-4.62,10.2-47.77-7.35,4.62-10.2-.82-.12-4.62,10.2-47.77-7.35,7.87-17.4,14.04,2.16.2-.46-14.04-2.16,7.87-17.4,14.04,2.16.2-.46-14.04-2.16,7.87-17.4,14.04,2.16.2-.46-14.04-2.16.65-1.43" fill="#fff" strokeWidth="0px"/>
+                </g>
+                </g>
+            </g>
+            <g id="long_stairs" data-name="long stairs">
+                <g>
+                <polygon points="1000.52 552.37 1000.46 555.11 997.3 562.11 997.35 559.37 1000.52 552.37" fill="#979797" strokeWidth="0px"/>
+                <polygon points="1383.38 611.25 1383.33 613.99 1000.46 555.11 1000.52 552.37 1383.38 611.25" fill="#959595" strokeWidth="0px"/>
+                <polygon points="1001.02 552.85 1000.97 555.59 998.14 561.84 998.19 559.1 1001.02 552.85" fill="#212121" strokeWidth="0px"/>
+                <polygon points="981.06 531.22 981.01 533.97 963.4 572.9 963.45 570.16 981.06 531.22" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1124.1 578.86 1124.04 581.6 997.3 562.11 997.35 559.37 1124.1 578.86" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1417.25 598.29 1417.2 601.04 981.01 533.97 981.06 531.22 1417.25 598.29" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1124.1 578.86 1124.04 581.6 1117.27 596.56 1117.33 593.83 1124.1 578.86" fill="#979797" strokeWidth="0px"/>
+                <polygon points="1117.33 593.83 1117.27 596.56 963.4 572.9 963.45 570.16 1117.33 593.83" fill="#959595" strokeWidth="0px"/>
+                <polygon points="1117.83 594.3 1117.77 597.03 962.56 573.17 962.61 570.43 1117.83 594.3" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1124.93 578.59 1124.88 581.32 1117.77 597.03 1117.83 594.3 1124.93 578.59" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1382.54 611.51 1382.49 614.26 1000.97 555.59 1001.02 552.85 1382.54 611.51" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1252.82 597.97 1252.76 600.72 1245.85 615.97 1245.91 613.24 1252.82 597.97" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1380.5 617.61 1380.45 620.35 1252.76 600.72 1252.82 597.97 1380.5 617.61" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1383.38 611.25 1383.33 613.99 1380.45 620.35 1380.5 617.61 1383.38 611.25" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1399.79 636.89 1399.73 639.63 1245.85 615.97 1245.91 613.24 1399.79 636.89" fill="#959595" strokeWidth="0px"/>
+                <polygon points="1417.25 598.29 1417.2 601.04 1399.73 639.63 1399.79 636.89 1417.25 598.29" fill="#979797" strokeWidth="0px"/>
+                <path d="M980.57,530.75l437.53,67.28-17.8,39.34-155.21-23.87,7.24-16,127.69,19.63,2.54-5.62-381.52-58.67-2.83,6.25,126.74,19.48-7.11,15.71-155.21-23.86,17.95-39.68v.02ZM1399.79,636.89l17.46-38.6-436.19-67.07-17.61,38.94,153.88,23.66,6.77-14.97-126.74-19.49,3.17-6.99,382.87,58.87-2.88,6.37-127.69-19.63-6.9,15.25,153.88,23.66" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="1417.25 598.29 1399.79 636.89 1245.91 613.24 1252.82 597.97 1380.5 617.61 1383.38 611.25 1000.52 552.37 997.35 559.37 1124.1 578.86 1117.33 593.83 963.45 570.16 981.06 531.22 1417.25 598.29" fill="#c7c7c7" strokeWidth="0px"/>
+                <polygon points="1418.1 598.03 1418.04 600.77 1400.24 640.12 1400.29 637.37 1418.1 598.03" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1400.29 637.37 1400.24 640.12 1245.02 616.24 1245.07 613.51 1400.29 637.37" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_1" data-name=" 1">
+                <g>
+                <polygon points="689.69 522.62 689.46 534.32 592.11 519.34 592.34 507.65 689.69 522.62" fill="#212121" strokeWidth="0px"/>
+                <polygon points="689.69 522.62 689.46 534.32 676.04 563.96 676.28 552.27 689.69 522.62" fill="#6d6d6d" strokeWidth="0px"/>
+                <polygon points="690.53 522.35 690.3 534.04 676.89 563.69 677.11 552 690.53 522.35" fill="#212121" strokeWidth="0px"/>
+                <polygon points="704.56 556.61 704.32 568.31 676.04 563.96 676.28 552.27 704.56 556.61" fill="#212121" strokeWidth="0px"/>
+                <polygon points="592.34 507.65 592.11 519.34 530.06 656.47 530.29 644.78 592.34 507.65" fill="#212121" strokeWidth="0px"/>
+                <polygon points="655.92 664.1 655.69 675.79 530.06 656.47 530.29 644.78 655.92 664.1" fill="#6c6c6c" strokeWidth="0px"/>
+                <polygon points="704.56 556.61 704.32 568.31 655.69 675.79 655.92 664.1 704.56 556.61" fill="#6d6d6d" strokeWidth="0px"/>
+                <path d="M677.11,552l28.28,4.35-48.97,108.23-126.97-19.53,62.39-137.88,98.69,15.17-13.41,29.64v.02ZM655.92,664.1l48.63-107.48-28.27-4.35,13.41-29.64-97.35-14.97-62.05,137.13,125.62,19.32" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="689.69 522.62 676.28 552.27 704.56 556.61 655.92 664.1 530.29 644.78 592.34 507.65 689.69 522.62" fill="#8f8f8f" strokeWidth="0px"/>
+                <polygon points="705.4 556.34 705.17 568.03 656.19 676.27 656.43 664.58 705.4 556.34" fill="#212121" strokeWidth="0px"/>
+                <polygon points="656.43 664.58 656.19 676.27 529.21 656.75 529.45 645.05 656.43 664.58" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_2" data-name=" 2">
+                <g>
+                <polygon points="803.89 572.2 803.66 583.89 705.36 568.78 705.59 557.08 803.89 572.2" fill="#212121" strokeWidth="0px"/>
+                <polygon points="705.59 557.08 705.36 568.78 656.73 676.26 656.96 664.57 705.59 557.08" fill="#212121" strokeWidth="0px"/>
+                <polygon points="755.25 679.69 755.02 691.38 656.73 676.26 656.96 664.57 755.25 679.69" fill="#787878" strokeWidth="0px"/>
+                <polygon points="803.89 572.2 803.66 583.89 755.02 691.38 755.25 679.69 803.89 572.2" fill="#7a7a7a" strokeWidth="0px"/>
+                <polygon points="803.89 572.2 755.25 679.69 656.96 664.57 705.59 557.08 803.89 572.2" fill="#a0a0a0" strokeWidth="0px"/>
+                <polygon points="755.75 680.16 755.52 691.85 655.88 676.53 656.12 664.84 755.75 680.16" fill="#212121" strokeWidth="0px"/>
+                <polygon points="804.73 571.92 804.5 583.62 755.52 691.85 755.75 680.16 804.73 571.92" fill="#212121" strokeWidth="0px"/>
+                <path d="M705.09,556.6l99.63,15.32-48.97,108.23-99.63-15.32,48.97-108.23h0ZM755.25,679.69l48.63-107.49-98.29-15.11-48.63,107.49,98.29,15.11" fill="#2c2c2c" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_3" data-name=" 3">
+                <g>
+                <polygon points="1033.22 607.31 1032.99 619 805.57 584.03 805.8 572.34 1033.22 607.31" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1033.22 607.31 1032.99 619 1032.79 619.44 1033.02 607.74 1033.22 607.31" fill="#575757" strokeWidth="0px"/>
+                <polygon points="1034.06 607.03 1033.83 618.73 1033.46 619.54 1033.69 607.85 1034.06 607.03" fill="#212121" strokeWidth="0px"/>
+                <polygon points="805.8 572.34 805.57 584.03 756.93 691.52 757.17 679.83 805.8 572.34" fill="#212121" strokeWidth="0px"/>
+                <polygon points="983.64 714.65 983.41 726.34 756.93 691.52 757.17 679.83 983.64 714.65" fill="#565656" strokeWidth="0px"/>
+                <polygon points="1033.02 607.74 1032.79 619.44 983.41 726.34 983.64 714.65 1033.02 607.74" fill="#575757" strokeWidth="0px"/>
+                <path d="M805.29,571.86l228.76,35.18-.37.81-49.56,107.28-227.81-35.03,48.97-108.23h0ZM1033.02,607.74l.2-.44-227.41-34.97-48.63,107.49,226.47,34.82,49.38-106.9" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="1033.22 607.31 1033.02 607.74 983.64 714.65 757.17 679.83 805.8 572.34 1033.22 607.31" fill="#727272" strokeWidth="0px"/>
+                <polygon points="984.14 715.13 983.91 726.82 756.09 691.79 756.32 680.1 984.14 715.13" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1033.69 607.85 1033.46 619.54 983.91 726.82 984.14 715.13 1033.69 607.85" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="pwd_stairs" data-name="pwd stairs">
+                <g>
+                <polygon points="1045.15 734.35 1045.12 735.99 936.3 719.26 936.33 717.62 1045.15 734.35" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1037.9 750.38 1037.86 752.02 951.94 738.81 951.97 737.17 1037.9 750.38" fill="#959595" strokeWidth="0px"/>
+                <polygon points="952.48 737.64 952.45 739.29 951.9 740.5 951.93 738.85 952.48 737.64" fill="#212121" strokeWidth="0px"/>
+                <polygon points="951.97 737.17 951.94 738.81 951.05 740.76 951.09 739.12 951.97 737.17" fill="#979797" strokeWidth="0px"/>
+                <polygon points="1045.15 734.35 1045.12 735.99 1037.86 752.02 1037.9 750.38 1045.15 734.35" fill="#979797" strokeWidth="0px"/>
+                <polygon points="1045.98 734.08 1045.95 735.73 1038.37 752.5 1038.39 750.85 1045.98 734.08" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1038.39 750.85 1038.37 752.5 952.45 739.29 952.48 737.64 1038.39 750.85" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1037.01 752.33 1036.98 753.98 951.05 740.76 951.09 739.12 1037.01 752.33" fill="#212121" strokeWidth="0px"/>
+                <polygon points="936.33 717.62 936.3 719.26 921.06 752.96 921.09 751.31 936.33 717.62" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1037.01 752.33 1036.98 753.98 1029.87 769.69 1029.91 768.05 1037.01 752.33" fill="#979797" strokeWidth="0px"/>
+                <polygon points="1029.91 768.05 1029.87 769.69 921.06 752.96 921.09 751.31 1029.91 768.05" fill="#959595" strokeWidth="0px"/>
+                <polygon points="1045.15 734.35 1037.9 750.38 951.97 737.17 951.09 739.12 1037.01 752.33 1029.91 768.05 921.09 751.31 936.33 717.62 1045.15 734.35" fill="#c7c7c7" strokeWidth="0px"/>
+                <path d="M1038.39,750.85l-85.92-13.21-.54,1.2,85.92,13.21-7.45,16.46-110.15-16.94,15.59-34.43,110.15,16.94-7.59,16.77h0ZM1037.01,752.33l-85.93-13.21.88-1.95,85.93,13.21,7.25-16.03-108.82-16.73-15.25,33.69,108.82,16.73,7.11-15.71" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="1030.4 768.52 1030.37 770.16 920.22 753.23 920.25 751.59 1030.4 768.52" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1037.85 752.07 1037.81 753.71 1030.37 770.16 1030.4 768.52 1037.85 752.07" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_4" data-name=" 4">
+                <g>
+                <polygon points="1263.13 730.09 1262.91 741.33 1202.39 732.02 1202.61 720.79 1263.13 730.09" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1202.61 720.79 1202.39 732.02 1190.62 758.02 1190.84 746.78 1202.61 720.79" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1251.37 756.09 1251.15 767.33 1190.62 758.02 1190.84 746.78 1251.37 756.09" fill="#626262" strokeWidth="0px"/>
+                <polygon points="1263.13 730.09 1262.91 741.33 1251.15 767.33 1251.37 756.09 1263.13 730.09" fill="#636363" strokeWidth="0px"/>
+                <polygon points="1263.13 730.09 1251.37 756.09 1190.84 746.78 1202.61 720.79 1263.13 730.09" fill="#828282" strokeWidth="0px"/>
+                <polygon points="1251.75 756.45 1251.53 767.68 1189.99 758.22 1190.22 746.99 1251.75 756.45" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1263.77 729.89 1263.55 741.12 1251.53 767.68 1251.75 756.45 1263.77 729.89" fill="#212121" strokeWidth="0px"/>
+                <path d="M1202.23,720.43l61.53,9.47-12.02,26.56-61.53-9.47,12.02-26.56ZM1251.37,756.09l11.76-26-60.52-9.31-11.76,26,60.52,9.31" fill="#2c2c2c" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_5" data-name=" 5">
+                <g>
+                <polygon points="1433.44 667.95 1433.21 679.64 1300.98 659.32 1301.21 647.62 1433.44 667.95" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1301.21 647.62 1300.98 659.32 1251.92 767.75 1252.16 756.06 1301.21 647.62" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1384.37 776.39 1384.14 788.08 1251.92 767.75 1252.16 756.06 1384.37 776.39" fill="#7c7c7c" strokeWidth="0px"/>
+                <polygon points="1433.44 667.95 1433.21 679.64 1384.14 788.08 1384.37 776.39 1433.44 667.95" fill="#7d7d7d" strokeWidth="0px"/>
+                <polygon points="1433.44 667.95 1384.37 776.39 1252.16 756.06 1301.21 647.62 1433.44 667.95" fill="#a4a4a4" strokeWidth="0px"/>
+                <polygon points="1384.88 776.87 1384.64 788.56 1251.08 768.02 1251.31 756.33 1384.88 776.87" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1434.28 667.69 1434.05 679.38 1384.64 788.56 1384.88 776.87 1434.28 667.69" fill="#212121" strokeWidth="0px"/>
+                <path d="M1300.71,647.15l133.57,20.54-49.41,109.18-133.57-20.54,49.41-109.18h0ZM1384.37,776.39l49.07-108.43-132.22-20.33-49.07,108.43,132.22,20.33" fill="#2c2c2c" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_6" data-name=" 6">
+                <g>
+                <polygon points="1560.36 687.99 1560.13 699.68 1434.51 680.36 1434.74 668.67 1560.36 687.99" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1434.74 668.67 1434.51 680.36 1385.73 788.16 1385.96 776.47 1434.74 668.67" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1511.58 795.78 1511.35 807.48 1385.73 788.16 1385.96 776.47 1511.58 795.78" fill="#717171" strokeWidth="0px"/>
+                <polygon points="1560.36 687.99 1560.13 699.68 1511.35 807.48 1511.58 795.78 1560.36 687.99" fill="#727272" strokeWidth="0px"/>
+                <path d="M1434.23,668.19l126.97,19.53-49.12,108.54-126.97-19.53,49.12-108.54ZM1511.58,795.78l48.78-107.8-125.62-19.32-48.78,107.8,125.62,19.32" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="1560.36 687.99 1511.58 795.78 1385.96 776.47 1434.74 668.67 1560.36 687.99" fill="#969696" strokeWidth="0px"/>
+                <polygon points="1512.09 796.26 1511.85 807.95 1384.88 788.43 1385.12 776.74 1512.09 796.26" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1561.21 687.72 1560.97 699.41 1511.85 807.95 1512.09 796.26 1561.21 687.72" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="_7" data-name=" 7">
+                <g>
+                <polygon points="1769.05 688.59 1768.82 700.28 1669.59 685.02 1669.82 673.33 1769.05 688.59" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1669.82 673.33 1669.59 685.02 1656.46 714.04 1656.69 702.34 1669.82 673.33" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1656.69 702.34 1656.46 714.04 1561.26 699.4 1561.5 687.71 1656.69 702.34" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1561.5 687.71 1561.26 699.4 1512.34 807.52 1512.58 795.82 1561.5 687.71" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1707 825.72 1706.77 837.42 1512.34 807.52 1512.58 795.82 1707 825.72" fill="#949494" strokeWidth="0px"/>
+                <polygon points="1769.05 688.59 1768.82 700.28 1706.77 837.42 1707 825.72 1769.05 688.59" fill="#959595" strokeWidth="0px"/>
+                <path d="M1669.32,672.86l100.58,15.47-62.38,137.87-195.77-30.1,49.26-108.87,95.19,14.64,13.13-29.02h0ZM1707,825.72l62.05-137.13-99.23-15.26-13.13,29.02-95.19-14.64-48.92,108.12,194.43,29.9" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="1769.05 688.59 1707 825.72 1512.58 795.82 1561.5 687.71 1656.69 702.34 1669.82 673.33 1769.05 688.59" fill="#c4c4c4" strokeWidth="0px"/>
+                <polygon points="1707.51 826.2 1707.27 837.89 1511.51 807.79 1511.74 796.1 1707.51 826.2" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1769.89 688.33 1769.66 700.02 1707.27 837.89 1707.51 826.2 1769.89 688.33" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="pi1">
+                <g>
+                <polygon points="742.48 523.91 742.21 537.62 735.12 536.53 735.4 522.82 742.48 523.91" fill="#212121" strokeWidth="0px"/>
+                <polygon points="735.4 522.82 735.12 536.53 733.34 540.48 733.61 526.77 735.4 522.82" fill="#212121" strokeWidth="0px"/>
+                <polygon points="740.69 527.86 740.42 541.56 733.34 540.48 733.61 526.77 740.69 527.86" fill="#2f2f2f" strokeWidth="0px"/>
+                <polygon points="742.48 523.91 742.21 537.62 740.42 541.56 740.69 527.86 742.48 523.91" fill="#303030" strokeWidth="0px"/>
+                <path d="M734.89,522.34l8.43,1.29-2.12,4.69-8.43-1.29,2.12-4.69ZM740.69,527.86l1.78-3.95-7.08-1.09-1.78,3.95,7.08,1.09" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="742.48 523.91 740.69 527.86 733.61 526.77 735.4 522.82 742.48 523.91" fill="#3f3f3f" strokeWidth="0px"/>
+                <polygon points="741.2 528.34 740.93 542.03 732.5 540.74 732.77 527.04 741.2 528.34" fill="#212121" strokeWidth="0px"/>
+                <polygon points="743.33 523.64 743.06 537.34 740.93 542.03 741.2 528.34 743.33 523.64" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="pi3">
+                <g>
+                <polygon points="848.67 542 848.4 555.71 841.31 554.61 841.58 540.91 848.67 542" fill="#212121" strokeWidth="0px"/>
+                <polygon points="841.58 540.91 841.31 554.61 839.52 558.56 839.79 544.85 841.58 540.91" fill="#212121" strokeWidth="0px"/>
+                <polygon points="846.89 545.95 846.61 559.65 839.52 558.56 839.79 544.85 846.89 545.95" fill="#2f2f2f" strokeWidth="0px"/>
+                <polygon points="848.67 542 848.4 555.71 846.61 559.65 846.89 545.95 848.67 542" fill="#303030" strokeWidth="0px"/>
+                <polygon points="848.67 542 846.89 545.95 839.79 544.85 841.58 540.91 848.67 542" fill="#3f3f3f" strokeWidth="0px"/>
+                <polygon points="847.39 546.43 847.11 560.13 838.68 558.83 838.95 545.13 847.39 546.43" fill="#212121" strokeWidth="0px"/>
+                <polygon points="849.51 541.73 849.24 555.43 847.11 560.13 847.39 546.43 849.51 541.73" fill="#212121" strokeWidth="0px"/>
+                <path d="M841.08,540.44l8.43,1.29-2.12,4.7-8.43-1.29,2.12-4.7ZM846.89,545.95l1.78-3.95-7.09-1.09-1.78,3.95,7.09,1.09" fill="#2c2c2c" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="pi11">
+                <g>
+                <polygon points="1521.02 643.41 1520.75 657.12 1513.65 656.03 1513.93 642.33 1521.02 643.41" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1513.93 642.33 1513.65 656.03 1511.87 659.98 1512.14 646.27 1513.93 642.33" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1519.23 647.36 1518.95 661.07 1511.87 659.98 1512.14 646.27 1519.23 647.36" fill="#2f2f2f" strokeWidth="0px"/>
+                <polygon points="1521.02 643.41 1520.75 657.12 1518.95 661.07 1519.23 647.36 1521.02 643.41" fill="#303030" strokeWidth="0px"/>
+                <path d="M1513.43,641.86l8.43,1.29-2.12,4.69-8.43-1.29,2.12-4.69ZM1519.23,647.36l1.78-3.95-7.08-1.09-1.78,3.95,7.08,1.09" fill="#2c2c2c" strokeWidth="0px"/>
+                <polygon points="1521.02 643.41 1519.23 647.36 1512.14 646.27 1513.93 642.33 1521.02 643.41" fill="#3f3f3f" strokeWidth="0px"/>
+                <polygon points="1519.74 647.84 1519.46 661.55 1511.03 660.25 1511.31 646.55 1519.74 647.84" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1521.86 643.15 1521.59 656.85 1519.46 661.55 1519.74 647.84 1521.86 643.15" fill="#212121" strokeWidth="0px"/>
+                </g>
+            </g>
+            <g id="pi12">
+                <g>
+                <polygon points="1632.57 661.66 1632.3 675.36 1625.21 674.27 1625.48 660.56 1632.57 661.66" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1625.48 660.56 1625.21 674.27 1623.42 678.22 1623.7 664.52 1625.48 660.56" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1630.78 665.61 1630.51 679.31 1623.42 678.22 1623.7 664.52 1630.78 665.61" fill="#2f2f2f" strokeWidth="0px"/>
+                <polygon points="1632.57 661.66 1632.3 675.36 1630.51 679.31 1630.78 665.61 1632.57 661.66" fill="#303030" strokeWidth="0px"/>
+                <polygon points="1632.57 661.66 1630.78 665.61 1623.7 664.52 1625.48 660.57 1632.57 661.66" fill="#3f3f3f" strokeWidth="0px"/>
+                <polygon points="1631.28 666.08 1631 679.78 1622.58 678.49 1622.85 664.79 1631.28 666.08" fill="#212121" strokeWidth="0px"/>
+                <polygon points="1633.41 661.39 1633.13 675.09 1631 679.78 1631.28 666.08 1633.41 661.39" fill="#212121" strokeWidth="0px"/>
+                <path d="M1624.97,660.09l8.43,1.29-2.12,4.69-8.43-1.29,2.12-4.69ZM1630.78,665.61l1.79-3.95-7.08-1.09-1.79,3.95,7.08,1.09" fill="#2c2c2c" strokeWidth="0px"/>
+                </g>
+            </g>
+            </g>
+        </g>
+        <g id="Layer_2" data-name="Layer 2">
+            <g id="floor-2" data-name="floor">
+            <g isolation="isolate">
+                <polygon points="991.86 519.07 991.59 533.25 982.6 553.71 982.88 539.53 991.86 519.07" fill="#767676" strokeWidth="0px"/>
+                <polygon points="991 523.5 990.73 537.68 989.72 537.52 989.99 523.34 991 523.5" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1373.81 579.56 1373.54 593.74 991.59 533.25 991.86 519.07 1373.81 579.56" fill="#747474" strokeWidth="0px"/>
+                <polygon points="992.62 519.81 992.35 533.99 990.73 537.68 991 523.5 992.62 519.81" fill="#291455" strokeWidth="0px"/>
+                <polygon points="971.27 498.19 971 512.37 954.78 549.29 955.06 535.11 971.27 498.19" fill="#291455" strokeWidth="0px"/>
+                <polygon points="955.06 535.11 954.78 549.29 585.37 490.54 585.64 476.36 955.06 535.11" fill="#281454" strokeWidth="0px"/>
+                <polygon points="990.75 524.07 990.48 538.25 983.71 553.65 983.98 539.47 990.75 524.07" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1372.55 579.98 1372.27 594.16 992.35 533.99 992.62 519.81 1372.55 579.98" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1116.25 543.94 1115.97 558.12 990.48 538.25 990.75 524.07 1116.25 543.94" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1116.25 543.94 1115.97 558.12 1109.21 573.53 1109.48 559.35 1116.25 543.94" fill="#646464" strokeWidth="0px"/>
+                <polygon points="1117.51 543.53 1117.24 557.71 1110.34 573.41 1110.62 559.23 1117.51 543.53" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1116.25 543.94 1109.48 559.35 983.98 539.48 990.75 524.07 1116.25 543.94" fill="#848484" strokeWidth="0px"/>
+                <polygon points="1109.48 559.35 1109.21 573.53 983.71 553.66 983.98 539.48 1109.48 559.35" fill="#636363" strokeWidth="0px"/>
+                <polygon points="1110.62 559.23 1110.34 573.41 1110.09 573.98 1110.36 559.8 1110.62 559.23" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1110.24 560.09 1109.96 574.27 982.44 554.08 982.72 539.89 1110.24 560.09" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1110.36 559.8 1110.09 573.98 1109.96 574.27 1110.24 560.09 1110.36 559.8" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1408.89 567.5 1408.61 581.68 971 512.37 971.27 498.19 1408.89 567.5" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1245.58 564.25 1245.3 578.43 1238.54 593.83 1238.81 579.65 1245.58 564.25" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1237.56 580.03 1237.29 594.21 1110.09 573.98 1110.36 559.8 1237.56 580.03" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1237.82 579.45 1237.54 593.63 1237.29 594.21 1237.56 580.03 1237.82 579.45" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1110.36 559.8 1110.62 559.23 1237.82 579.45 1237.56 580.03 1110.36 559.8" fill="#361b70" strokeWidth="0px"/>
+                <polygon points="1372.02 583.65 1371.75 597.83 1370.73 597.67 1371.01 583.48 1372.02 583.65" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1373.81 579.56 1373.54 593.74 1371.75 597.83 1372.02 583.65 1373.81 579.56" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1371.08 584.12 1370.8 598.3 1245.3 578.43 1245.58 564.25 1371.08 584.12" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1364.31 599.53 1364.03 613.71 1238.54 593.83 1238.81 579.65 1364.31 599.53" fill="#636363" strokeWidth="0px"/>
+                <polygon points="1290.02 588.37 1289.74 602.55 1237.27 594.24 1237.55 580.06 1290.02 588.37" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1371.08 584.12 1370.8 598.3 1364.03 613.71 1364.31 599.53 1371.08 584.12" fill="#646464" strokeWidth="0px"/>
+                <polygon points="1371.08 584.12 1364.31 599.53 1238.81 579.65 1245.58 564.25 1371.08 584.12" fill="#848484" strokeWidth="0px"/>
+                <polygon points="1364.74 600.21 1364.47 614.39 1289.74 602.55 1290.02 588.37 1364.74 600.21" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1364.74 600.21 1364.47 614.39 1289.74 602.55 1290.02 588.37 1364.74 600.21" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1365.06 600.26 1364.78 614.44 1364.47 614.39 1364.74 600.21 1365.06 600.26" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1372.33 583.69 1372.06 597.88 1364.78 614.44 1365.06 600.26 1372.33 583.69" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1408.89 567.5 1408.61 581.68 1392.28 618.86 1392.55 604.68 1408.89 567.5" fill="#767676" strokeWidth="0px"/>
+                <polygon points="1410.14 567.08 1409.87 581.26 1393.54 618.44 1393.81 604.26 1410.14 567.08" fill="#291455" strokeWidth="0px"/>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-6)">
+                    <g isolation="isolate">
+                    <path d="M585.64,476.36l-.27,14.18c-15.66,34.46-42.08,93.1-52.52,117.86l.27-14.18c10.44-24.76,36.85-83.4,52.52-117.86" fill="#281453" strokeWidth="0px"/>
+                    <path d="M533.12,594.22l-.27,14.18c-2.36,5.59-3.9,9.45-4.32,10.92l.27-14.18c.42-1.47,1.97-5.33,4.32-10.92" fill="#261350" strokeWidth="0px"/>
+                    <path d="M528.8,605.14l-.27,14.18c-.06.19-.09.35-.11.46l.27-14.18c.02-.11.05-.26.11-.46" fill="#25134d" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <polygon points="1764.25 663.79 1763.97 677.97 1392.28 618.86 1392.55 604.68 1764.25 663.79" fill="#281454" strokeWidth="0px"/>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-7)">
+                    <g isolation="isolate">
+                    <path d="M1707.78,792.33l-.27,14.18c-13.95-2.23-302.94-48.29-588.57-93.68-117.87-18.73-212.21-33.7-287.73-45.66-59.64-9.45-107.54-17.02-146.02-23.08l.27-14.18c38.48,6.07,86.39,13.64,146.02,23.08,75.53,11.96,169.87,26.93,287.73,45.66,285.63,45.39,574.62,91.46,588.57,93.68" fill="#686868" strokeWidth="0px"/>
+                    <path d="M685.45,629.91l-.27,14.18c-132.5-20.89-153.29-23.94-156.76-24.31l.27-14.18c3.47.37,24.26,3.42,156.76,24.31" fill="#6b6b6b" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-8)">
+                    <g isolation="isolate">
+                    <path d="M1707.78,792.33l-.27,14.18c-13.95-2.23-302.94-48.29-588.57-93.68-117.87-18.73-212.21-33.7-287.73-45.66-59.64-9.45-107.54-17.02-146.02-23.08l.27-14.18c38.48,6.07,86.39,13.64,146.02,23.08,75.53,11.96,169.87,26.93,287.73,45.66,285.63,45.39,574.62,91.46,588.57,93.68" fill="#686868" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <path d="M1392.55,604.68l371.69,59.11-56.47,128.54c-13.95-2.23-302.94-48.29-588.57-93.68-117.87-18.73-212.21-33.7-287.73-45.66-264.99-41.97-298.31-46.92-302.78-47.39.69-4.63,37.29-85.98,56.95-129.24l369.42,58.75,16.22-36.92,437.61,69.31-16.33,37.18ZM1364.72,600.26l7.61-16.56,1.48-4.13-381.95-60.49-8.99,20.46,127.36,20.56,127.32,19.94,52.45,8.34,74.71,11.88" fill="#9b9b9b" strokeWidth="0px"/>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-9)">
+                    <g isolation="isolate">
+                    <polyline points="528.32 605.54 528.05 619.72 528.05 619.72 528.32 605.54 528.32 605.54" fill="#24124b" strokeWidth="0px"/>
+                    <path d="M528.32,605.54l-.27,14.18s-.04,0-.05,0l.27-14.18s.04,0,.05,0" fill="#25134d" strokeWidth="0px"/>
+                    <path d="M528.27,605.53l-.27,14.18s-.04,0-.06,0l.27-14.18s.04,0,.06,0" fill="#261350" strokeWidth="0px"/>
+                    <path d="M528.21,605.53l-.27,14.18s-.02,0-.03,0l.27-14.18s.02,0,.03,0" fill="#281453" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <polygon points="528.42 605.56 528.14 619.74 528.05 619.72 528.32 605.54 528.42 605.56" fill="#271451" strokeWidth="0px"/>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-10)">
+                    <g isolation="isolate">
+                    <polyline points="528.18 605.53 527.91 619.71 527.9 619.71 528.17 605.53 528.18 605.53" fill="#281453" strokeWidth="0px"/>
+                    <path d="M528.17,605.53l-.27,14.18s-.04,0-.06,0l.27-14.18s.04,0,.06,0" fill="#291455" strokeWidth="0px"/>
+                    <path d="M528.11,605.53l-.27,14.18s-.04,0-.06,0l.27-14.18s.04,0,.06,0" fill="#2a1558" strokeWidth="0px"/>
+                    <path d="M528.05,605.54l-.27,14.18s-.05.01-.08.02l.27-14.18s.05-.01.08-.02" fill="#2b165a" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-11)">
+                    <g isolation="isolate">
+                    <path d="M528.41,605.57l-.27,14.18c-.1,0-.18-.02-.24-.02l.27-14.18c.06,0,.14,0,.24.02" fill="#6f6f6f" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-12)">
+                    <g isolation="isolate">
+                    <path d="M528.38,606.13l-.27,14.18s-.01,0-.02,0l.27-14.18s.01,0,.02,0" fill="#231148" strokeWidth="0px"/>
+                    <path d="M528.37,606.12l-.27,14.18s-.06-.02-.09-.03l.27-14.18s.06.02.09.03" fill="#211146" strokeWidth="0px"/>
+                    <path d="M528.27,606.09l-.27,14.18c-.14-.06-.26-.16-.25-.53l.27-14.18c0,.36.11.47.25.53" fill="#201043" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-13)">
+                    <g isolation="isolate">
+                    <path d="M528.41,605.57l-.27,14.18c-.1,0-.18-.02-.24-.02l.27-14.18c.06,0,.14,0,.24.02" fill="#6f6f6f" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-14)">
+                    <g isolation="isolate">
+                    <path d="M528.41,605.57l-.27,14.18c-.1,0-.18-.02-.24-.02l.27-14.18c.06,0,.14,0,.24.02" fill="#6f6f6f" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <path d="M528.18,605.53s.1,0,.14.01l.09.02h0c-.1,0-.18,0-.24,0,0,0,0-.02,0-.02Z" fill="#9b9b9b" strokeWidth="0px"/>
+                <polygon points="1764.25 663.79 1763.97 677.97 1707.5 806.52 1707.78 792.33 1764.25 663.79" fill="#767676" strokeWidth="0px"/>
+                <path d="M1393.81,604.26l371.7,59.11-56.98,129.71c-8.33-1.33-528.57-84.27-875.54-139.22-170.69-27.04-299.38-47.28-304.62-47.73-1.04-.3-1.8-.62,27.23-65.56,14.3-32,29-64.33,29.14-64.66l369.54,58.46,16.22-36.92,439.63,69.63-16.33,37.18ZM1707.78,792.34l56.47-128.55-371.69-59.11,16.33-37.18-437.61-69.31-16.22,36.92-369.42-58.75c-19.66,43.26-56.26,124.61-56.95,129.24,4.47.48,37.79,5.42,302.78,47.4,75.53,11.96,169.87,26.93,287.73,45.66,285.63,45.39,574.62,91.46,588.57,93.68M528.41,605.57h0s-.09-.03-.09-.03c-.05,0-.1-.01-.14-.01-.07,0-.14,0-.21.03.03,0,.08-.01.2,0,.06,0,.14,0,.24.02" fill="#361b70" strokeWidth="0px"/>
+                <path d="M991,523.5l126.51,20.03-6.9,15.7-.25.58-.13.29-127.52-20.2.16-.36,7.11-16.19,1.01.16ZM1109.48,559.35l6.77-15.41-125.5-19.87-6.77,15.4,125.5,19.88" fill="#361b70" strokeWidth="0px"/>
+                <polygon points="989.99 523.34 991.87 519.07 1373.81 579.56 1372.02 583.65 1371.01 583.48 1372.55 579.98 992.62 519.81 991 523.5 989.99 523.34" fill="#361b70" strokeWidth="0px"/>
+                <path d="M1372.02,583.65l.31.05-7.28,16.56-.31-.05-74.73-11.84-52.47-8.31.27-.61,7.01-15.95,126.19,19.98,1.01.16ZM1364.31,599.52l6.77-15.41-125.5-19.87-6.77,15.4,125.5,19.88" fill="#361b70" strokeWidth="0px"/>
+                <g isolation="isolate">
+                <g clip-path="url(#clippath-15)">
+                    <g isolation="isolate">
+                    <path d="M1708.54,793.08l-.27,14.18c-8.33-1.33-528.57-84.27-875.54-139.22-55.47-8.78-106.5-16.85-150.3-23.76l.27-14.18c43.8,6.9,94.83,14.97,150.3,23.76,346.97,54.95,867.21,137.89,875.54,139.22" fill="#24124b" strokeWidth="0px"/>
+                    <path d="M682.7,630.1l-.27,14.18c-89.32-14.08-148.57-23.32-154.07-23.94l.27-14.18c5.5.62,64.75,9.86,154.07,23.94" fill="#25134d" strokeWidth="0px"/>
+                    <path d="M528.63,606.15l-.27,14.18c-.1-.01-.19-.02-.25-.03l.27-14.18c.07,0,.15.01.25.03" fill="#261350" strokeWidth="0px"/>
+                    </g>
+                </g>
+                </g>
+                <polygon points="1765.51 663.37 1765.24 677.55 1708.26 807.26 1708.54 793.08 1765.51 663.37" fill="#291455" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g opacity=".17">
+            <g isolation="isolate">
+                <g clip-path="url(#clippath-16)">
+                <g clip-path="url(#clippath-17)">
+                    <path d="M577.53,597.17l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M634.73,606.23l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M691.94,615.29l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M749.15,624.35l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M806.35,633.41l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M863.55,642.47l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M920.76,651.53l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M977.96,660.59l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1035.16,669.65l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1092.37,678.71l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1149.58,687.77l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1206.78,696.83l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1263.98,705.89l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1321.19,714.95l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1378.39,724.01l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1435.59,733.07l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1492.8,742.13l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1550.01,751.19l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1607.21,760.25l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M587.04,575.51l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M644.25,584.57l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M701.45,593.64l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M758.66,602.7l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M815.86,611.76l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M873.06,620.82l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M930.27,629.88l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M987.47,638.94l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1044.68,648l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1101.88,657.06l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1159.09,666.12l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1216.29,675.18l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1273.49,684.24l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1330.7,693.3l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1387.9,702.36l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1445.11,711.42l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1502.31,720.48l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1559.52,729.54l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1616.72,738.6l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M596.55,553.86l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M653.76,562.92l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M710.97,571.98l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M768.17,581.04l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M825.37,590.1l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M882.58,599.16l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M939.79,608.22l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M996.98,617.28l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1054.19,626.34l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1111.4,635.4l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1168.6,644.46l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1225.8,653.52l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1283.01,662.58l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1340.22,671.64l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1397.41,680.7l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1454.62,689.76l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1511.83,698.82l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1569.03,707.88l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1626.23,716.94l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M606.06,532.2l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M663.27,541.26l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M720.48,550.32l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M777.69,559.39l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M834.88,568.44l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M892.09,577.51l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M949.3,586.57l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1006.5,595.63l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1063.7,604.69l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1120.91,613.75l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1178.12,622.81l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1235.31,631.87l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1292.52,640.93l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1349.73,649.99l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1406.93,659.05l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1464.13,668.11l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1521.34,677.17l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1578.55,686.23l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1635.74,695.29l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M615.58,510.55l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M672.79,519.61l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M729.99,528.67l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M787.2,537.73l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M844.4,546.79l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M901.6,555.85l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M958.81,564.91l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1016.01,573.97l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1073.22,583.03l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1130.42,592.09l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1187.63,601.15l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1244.83,610.21l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1302.03,619.27l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1359.24,628.33l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1416.44,637.39l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1473.65,646.45l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1530.85,655.51l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1588.06,664.57l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1645.26,673.63l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M968.32,543.26l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1025.52,552.31l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1082.73,561.38l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1139.93,570.44l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1197.14,579.5l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1254.34,588.56l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1311.55,597.62l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M977.83,521.6l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1035.03,530.66l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1092.24,539.72l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1149.45,548.78l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1206.65,557.84l9.27-21.1,56.23,8.91-9.27,21.1-56.23-8.91M1263.85,566.9l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1321.06,575.96l9.27-21.1,56.24,8.91-9.27,21.1-56.24-8.91M1388.43,561.87l-.97-.15-.66,1.49-56.24-8.91.66-1.49-.97-.15-.66,1.49-56.24-8.91.66-1.49-.97-.15-.66,1.49-56.23-8.91.66-1.49-.97-.15-.66,1.49-56.24-8.91.66-1.49-.97-.15-.66,1.49-56.24-8.91.66-1.49-.97-.15-.66,1.49-56.24-8.91.66-1.49-.97-.15-.66,1.49-56.23-8.91.66-1.49-.97-.15-.66,1.49-15.27-2.42-.24.56,15.27,2.42-9.27,21.1-15.27-2.42-.24.56,15.27,2.42-9.27,21.1-56.24-8.91,3.47-7.91-.97-.15-3.47,7.91-56.24-8.91,3.49-7.94-.97-.15-3.49,7.94-56.23-8.91,3.5-7.97-.97-.15-3.5,7.98-56.24-8.91,3.52-8.01-.97-.15-3.52,8.01-56.24-8.91,3.53-8.04-.97-.15-3.53,8.04-56.24-8.91,3.55-8.08-.96-.15-3.55,8.08-41.66-6.6c-.08.18-.17.37-.25.55l41.67,6.6-9.27,21.1-41.92-6.64c-.08.18-.17.37-.25.55l41.93,6.64-9.27,21.1-42.14-6.67c-.08.18-.17.37-.25.56l42.15,6.68-9.27,21.1-42.3-6.7c-.08.19-.16.37-.25.56l42.31,6.7-9.27,21.1-42.38-6.71c-.08.18-.16.37-.24.55l42.38,6.71-9.27,21.1-42.28-6.7c-.08.19-.16.37-.24.56l42.27,6.69-5.53,12.6c.32.05.64.1.96.15l5.53-12.59,56.24,8.91-5.48,12.48c.32.05.65.1.97.15l5.48-12.48,56.24,8.91-5.46,12.43c.33.05.64.1.97.15l5.46-12.43,56.24,8.91-5.45,12.4c.34.05.63.1.97.15l5.45-12.4,56.23,8.91-5.44,12.39c.34.05.63.1.97.15l5.44-12.39,56.24,8.91-5.44,12.39c.33.05.64.1.97.15l5.44-12.39,56.24,8.91-5.44,12.39c.34.05.63.1.97.15l5.44-12.39,56.23,8.91-5.45,12.41c.31.05.65.1.97.15l5.45-12.41,56.24,8.91-5.46,12.43c.32.05.65.1.97.15l5.46-12.43,56.24,8.91-5.47,12.45c.32.05.65.1.97.15l5.47-12.45,56.24,8.91-5.48,12.48c.32.05.64.1.97.15l5.48-12.48,56.23,8.91-5.5,12.51c.32.05.65.1.97.15l5.5-12.51,56.24,8.91-5.51,12.54c.32.05.65.1.97.15l5.51-12.55,56.24,8.91-5.53,12.58c.34.05.63.1.97.15l5.53-12.58,56.24,8.91-5.54,12.62c.33.05.63.1.96.15l5.54-12.62,56.24,8.91-5.56,12.66c.34.05.63.1.97.15l5.56-12.66,56.24,8.91-5.58,12.71c.32.05.65.1.97.15l5.58-12.71,56.24,8.91-5.6,12.76c.34.05.63.1.97.15l5.6-12.76,56.23,8.91-5.62,12.8c.33.05.64.1.97.15l5.63-12.81,56.24,8.91-5.65,12.86c.32.05.65.1.97.15l5.65-12.86,50.52,8,.24-.56-50.52-8,9.27-21.1,50.52,8,.24-.55-50.52-8,9.27-21.1,50.52,8,.24-.56-50.52-8,9.27-21.1,50.52,8,.24-.56-50.52-8,9.27-21.1,50.52,8,.24-.56-50.52-8,9.27-21.1,50.52,8,.24-.56-50.52-8,3.26-7.43-.97-.15-3.26,7.43-56.24-8.91,3.28-7.46-.97-.15-3.28,7.46-56.23-8.91,3.29-7.5-.97-.15-3.29,7.5-56.24-8.91,3.31-7.53-.97-.15-3.31,7.53-56.24-8.91,3.32-7.57-.97-.15-3.32,7.57-56.24-8.91,3.34-7.6-.96-.15-3.34,7.6-56.24-8.91,9.27-21.1,21.95,3.48.24-.56-21.95-3.48,9.27-21.1,21.95,3.48.24-.56-21.95-3.48.66-1.49" fill="#fff" strokeWidth="0px"/>
+                </g>
+                </g>
+            </g>
+            </g>
+            <g id="_1-2" data-name="1">
+            <g isolation="isolate">
+                <polygon points="681.89 486.59 681.55 504.01 580.94 488.07 581.28 470.65 681.89 486.59" fill="#281454" strokeWidth="0px"/>
+                <polygon points="681.89 486.59 681.55 504.01 669.6 530.69 669.94 513.27 681.89 486.59" fill="#a8713f" strokeWidth="0px"/>
+                <polygon points="683.11 486.19 682.77 503.61 670.82 530.29 671.15 512.87 683.11 486.19" fill="#291455" strokeWidth="0px"/>
+                <polygon points="690.5 516.53 690.16 533.95 669.6 530.69 669.94 513.27 690.5 516.53" fill="#281454" strokeWidth="0px"/>
+                <polygon points="581.28 470.65 580.94 488.07 529.21 605.84 529.55 588.42 581.28 470.65" fill="#291455" strokeWidth="0px"/>
+                <polygon points="681.89 486.59 669.94 513.27 690.5 516.53 650.5 607.57 529.55 588.42 581.28 470.65 681.89 486.59" fill="#dc9452" strokeWidth="0px"/>
+                <polygon points="650.5 607.57 650.17 625 529.21 605.84 529.55 588.42 650.5 607.57" fill="#a66f3e" strokeWidth="0px"/>
+                <polygon points="690.5 516.53 690.16 533.95 650.17 625 650.5 607.57 690.5 516.53" fill="#a7703e" strokeWidth="0px"/>
+                <path d="M671.15,512.87l20.56,3.26-40.48,92.16-122.9-19.47,52.22-118.88,102.56,16.24-11.95,26.68ZM650.5,607.57l40-91.05-20.56-3.26,11.95-26.68-100.61-15.93-51.73,117.76,120.96,19.16" fill="#361b70" strokeWidth="0px"/>
+                <polygon points="691.71 516.12 691.38 533.54 650.89 625.7 651.23 608.28 691.71 516.12" fill="#291455" strokeWidth="0px"/>
+                <polygon points="651.23 608.28 650.89 625.7 527.99 606.24 528.32 588.82 651.23 608.28" fill="#281454" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_2-2" data-name="2">
+            <g isolation="isolate">
+                <polygon points="757.25 526.73 756.91 544.15 692.2 533.9 692.53 516.48 757.25 526.73" fill="#281454" strokeWidth="0px"/>
+                <polygon points="692.53 516.48 692.2 533.9 652.06 625.28 652.39 607.86 692.53 516.48" fill="#291455" strokeWidth="0px"/>
+                <polygon points="717.1 618.11 716.77 635.53 652.06 625.28 652.39 607.86 717.1 618.11" fill="#a65070" strokeWidth="0px"/>
+                <polygon points="757.25 526.73 756.91 544.15 716.77 635.53 717.1 618.11 757.25 526.73" fill="#a75171" strokeWidth="0px"/>
+                <polygon points="757.25 526.73 717.1 618.11 652.39 607.86 692.53 516.48 757.25 526.73" fill="#dc6a95" strokeWidth="0px"/>
+                <polygon points="717.83 618.82 717.49 636.24 650.84 625.68 651.18 608.26 717.83 618.82" fill="#281454" strokeWidth="0px"/>
+                <polygon points="758.46 526.33 758.12 543.75 717.49 636.24 717.83 618.82 758.46 526.33" fill="#291455" strokeWidth="0px"/>
+                <path d="M691.81,515.77l66.65,10.56-40.63,92.49-66.65-10.56,40.63-92.49ZM717.1,618.11l40.14-91.38-64.71-10.25-40.14,91.38,64.71,10.25" fill="#361b70" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_3-2" data-name="3">
+            <g isolation="isolate">
+                <polygon points="837.16 539.57 836.82 556.99 757.91 544.32 758.24 526.9 837.16 539.57" fill="#281454" strokeWidth="0px"/>
+                <polygon points="758.24 526.9 757.91 544.32 717.84 635.53 718.17 618.11 758.24 526.9" fill="#291455" strokeWidth="0px"/>
+                <polygon points="797.16 630.62 796.82 648.04 717.84 635.53 718.17 618.11 797.16 630.62" fill="#53a68a" strokeWidth="0px"/>
+                <polygon points="837.16 539.57 836.82 556.99 796.82 648.04 797.16 630.62 837.16 539.57" fill="#53a78c" strokeWidth="0px"/>
+                <polygon points="758.24 526.9 837.16 539.57 797.16 630.62 718.17 618.11 758.24 526.9" fill="#6edcb8" strokeWidth="0px"/>
+                <polygon points="797.89 631.33 797.55 648.75 716.63 635.93 716.96 618.51 797.89 631.33" fill="#281454" strokeWidth="0px"/>
+                <polygon points="838.37 539.17 838.03 556.59 797.55 648.75 797.89 631.33 838.37 539.17" fill="#291455" strokeWidth="0px"/>
+                <path d="M757.52,526.19l80.85,12.98-40.48,92.16-80.92-12.82,40.56-92.33ZM797.16,630.62l40-91.05-78.91-12.68-40.07,91.21,78.99,12.51" fill="#361b70" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_4-2" data-name="4">
+            <g isolation="isolate">
+                <polygon points="963.97 559.35 963.63 576.78 837.89 556.68 838.22 539.26 963.97 559.35" fill="#281454" strokeWidth="0px"/>
+                <polygon points="838.22 539.26 837.89 556.68 797.89 647.74 798.23 630.31 838.22 539.26" fill="#291455" strokeWidth="0px"/>
+                <polygon points="924.04 650.24 923.71 667.66 797.89 647.74 798.23 630.31 924.04 650.24" fill="#0095a6" strokeWidth="0px"/>
+                <polygon points="963.97 559.35 963.63 576.78 923.71 667.66 924.04 650.24 963.97 559.35" fill="#0096a7" strokeWidth="0px"/>
+                <path d="M837.5,538.55l127.69,20.4-40.41,92-127.76-20.24,40.49-92.16ZM924.04,650.24l39.93-90.89-125.75-20.09-40,91.05,125.82,19.93" fill="#361b70" strokeWidth="0px"/>
+                <polygon points="838.22 539.26 963.97 559.35 924.04 650.24 798.23 630.31 838.22 539.26" fill="#00c6dc" strokeWidth="0px"/>
+                <polygon points="924.78 650.95 924.44 668.37 796.68 648.14 797.01 630.72 924.78 650.95" fill="#281454" strokeWidth="0px"/>
+                <polygon points="965.19 558.95 964.86 576.37 924.44 668.37 924.78 650.95 965.19 558.95" fill="#291455" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_5-2" data-name="5">
+            <g isolation="isolate">
+                <polygon points="1088.4 579.11 1088.07 596.53 965.11 577.06 965.45 559.64 1088.4 579.11" fill="#281454" strokeWidth="0px"/>
+                <polygon points="965.45 559.64 965.11 577.06 925.19 667.95 925.52 650.53 965.45 559.64" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1048.48 670 1048.14 687.42 925.19 667.95 925.52 650.53 1048.48 670" fill="#a6443e" strokeWidth="0px"/>
+                <polygon points="1088.4 579.11 1088.07 596.53 1048.14 687.42 1048.48 670 1088.4 579.11" fill="#a7443e" strokeWidth="0px"/>
+                <polygon points="1088.4 579.11 1048.48 670 925.52 650.53 965.45 559.64 1088.4 579.11" fill="#dc5a52" strokeWidth="0px"/>
+                <polygon points="1049.2 670.71 1048.87 688.13 923.96 668.35 924.3 650.93 1049.2 670.71" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1089.62 578.71 1089.28 596.13 1048.87 688.13 1049.2 670.71 1089.62 578.71" fill="#291455" strokeWidth="0px"/>
+                <path d="M964.71,558.93l124.9,19.78-40.41,92-124.9-19.78,40.41-92ZM1048.48,670l39.93-90.89-122.96-19.47-39.93,90.89,122.96,19.47" fill="#361b70" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_7-2" data-name="7">
+            <g isolation="isolate">
+                <polygon points="1370.53 624.67 1370.19 642.1 1229.52 619.82 1229.86 602.39 1370.53 624.67" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1229.86 602.39 1229.52 619.82 1189.67 710.54 1190.01 693.12 1229.86 602.39" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1330.68 715.4 1330.34 732.82 1189.67 710.54 1190.01 693.12 1330.68 715.4" fill="#a64b88" strokeWidth="0px"/>
+                <polygon points="1370.53 624.67 1370.19 642.1 1330.34 732.82 1330.68 715.4 1370.53 624.67" fill="#a74c8a" strokeWidth="0px"/>
+                <polygon points="1370.53 624.67 1330.68 715.4 1190.01 693.12 1229.86 602.39 1370.53 624.67" fill="#dc64b6" strokeWidth="0px"/>
+                <polygon points="1331.4 716.11 1331.06 733.53 1188.45 710.95 1188.79 693.52 1331.4 716.11" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1371.74 624.27 1371.4 641.69 1331.06 733.53 1331.4 716.11 1371.74 624.27" fill="#291455" strokeWidth="0px"/>
+                <path d="M1229.13,601.68l142.61,22.59-40.34,91.84-142.61-22.59,40.34-91.84ZM1330.68,715.4l39.85-90.72-140.67-22.28-39.85,90.72,140.67,22.28" fill="#361b70" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_8" data-name="8">
+            <g isolation="isolate">
+                <polygon points="1629.09 665.65 1628.76 683.07 1372.15 642.43 1372.48 625.01 1629.09 665.65" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1372.48 625.01 1372.15 642.43 1332.29 733.15 1332.63 715.73 1372.48 625.01" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1589.24 756.38 1588.91 773.8 1332.29 733.15 1332.63 715.73 1589.24 756.38" fill="#a63e3e" strokeWidth="0px"/>
+                <polygon points="1629.09 665.65 1628.76 683.07 1588.91 773.8 1589.24 756.38 1629.09 665.65" fill="#a73e3e" strokeWidth="0px"/>
+                <polygon points="1629.09 665.65 1589.24 756.38 1332.63 715.73 1372.48 625.01 1629.09 665.65" fill="#dc5252" strokeWidth="0px"/>
+                <polygon points="1589.97 757.08 1589.63 774.51 1331.08 733.56 1331.42 716.13 1589.97 757.08" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1630.31 665.25 1629.97 682.67 1589.63 774.51 1589.97 757.08 1630.31 665.25" fill="#291455" strokeWidth="0px"/>
+                <path d="M1371.76,624.3l258.55,40.95-40.34,91.83-258.55-40.95,40.34-91.83ZM1589.24,756.38l39.85-90.72-256.61-40.64-39.85,90.72,256.61,40.64" fill="#361b70" strokeWidth="0px"/>
+            </g>
+            </g>
+            <g id="_9" data-name="9">
+            <g isolation="isolate">
+                <polygon points="1759.43 657.91 1759.1 675.33 1658.69 659.42 1659.02 642 1759.43 657.91" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1659.02 642 1658.69 659.42 1647.03 685.97 1647.36 668.55 1659.02 642" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1647.36 668.55 1647.03 685.97 1627.9 682.94 1628.24 665.52 1647.36 668.55" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1628.24 665.52 1627.9 682.94 1588.12 773.5 1588.46 756.08 1628.24 665.52" fill="#291455" strokeWidth="0px"/>
+                <polygon points="1707.99 775.01 1707.66 792.43 1588.12 773.5 1588.46 756.08 1707.99 775.01" fill="#5b8129" strokeWidth="0px"/>
+                <polygon points="1759.43 657.91 1759.1 675.33 1707.66 792.43 1707.99 775.01 1759.43 657.91" fill="#5c8229" strokeWidth="0px"/>
+                <path d="M1658.3,641.29l102.35,16.21-51.93,118.22-121.47-19.24,40.27-91.67,19.13,3.03,11.66-26.55ZM1707.99,775.01l51.44-117.1-100.41-15.9-11.66,26.55-19.13-3.03-39.78,90.56,119.54,18.93" fill="#361b70" strokeWidth="0px"/>
+                <polygon points="1759.43 657.91 1707.99 775.01 1588.46 756.08 1628.24 665.52 1647.36 668.55 1659.02 642 1759.43 657.91" fill="#79ab36" strokeWidth="0px"/>
+                <polygon points="1708.71 775.72 1708.38 793.14 1586.9 773.9 1587.24 756.48 1708.71 775.72" fill="#281454" strokeWidth="0px"/>
+                <polygon points="1760.65 657.5 1760.31 674.92 1708.38 793.14 1708.71 775.72 1760.65 657.5" fill="#291455" strokeWidth="0px"/>
+            </g>
+            </g>
+        </g>
+        </svg>
+    </div>
+  )
+}
+
+export default NewAdminFloor2
